### 2024-01-16 #6232 @editorialbot
[PRE REVIEW]: gsMAMS: an R package for Designing Multi-Arm Multi-Stage Clinical Trials
**Submitting author:** @Tpatni719 (<a href="http://orcid.org/0000-0003-1605-7207">Tushar Patni</a>)
**Repository:** https://github.com/Tpatni719/gsMAMS.git
**Branch with paper.md** (empty if default branch):
**Version:** 0.7.1
