### 2024-02-02 #6232 @njtierney
I’d be happy to review! But if you’ve me to be the editor I wouldn’t mind having another review for JOSS under my belt before I agreed to be the editor. When would you like my review by?
