### 2024-02-03 #6232 @ppxasjsm
Hi @RhysPeploe, would you be interested in reviewing this JOSS submission?
