### 2024-03-14 Tpatni719/gsMAMS#1 @RhysPeploe
Improved Documentation
https://github.com/openjournals/joss-reviews/issues/6322

The addition of:
- [x] a README.md file and / or a description in 'About'
- [x] directions on how to install the package in the paper and the GitHub. The paper referrences to [https://cran.r-project.org/web/packages/gsMAMS/index.html](https://cran.r-project.org/web/packages/gsMAMS/index.html) but for the sake of less experienced R users, it would be good to include the install.packages() and library() commands to improve remove neccassary barriers to using the package
- [x] Community guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support

would improve the package.
