### 2024-03-14 Tpatni719/gsMAMS#1 @Tpatni719
Thank you so much for the suggestions! I have added the recommendations in the package.
