### 2024-03-15 Tpatni719/gsMAMS#2 @RhysPeploe
Inconsistencies in displayed code in paper
https://github.com/openjournals/joss-reviews/issues/6322

Minor, but important for quality of the paper, some of the example code has spaces between items in the concatenated lists, and around the equals sign, then some code does not. Keeping it consistent would make the paper higher quality.

![image](https://github.com/Tpatni719/gsMAMS/assets/68282471/ea8fa644-8b1c-4f24-9a29-e97f608a0820)
