### 2024-03-15 Tpatni719/gsMAMS#3 @RhysPeploe
set.seed( ) & results of example code
https://github.com/openjournals/joss-reviews/issues/6322

My understanding is that these functions are deterministic, does the seed affect the results? If the seed does not imapct results, this can be removed?

It would be good to report the results you get from the code in the paper so that users can compare their results to those in the paper too
