### 2024-03-15 ed2f911
Made the syntax in the paper consistent.
