+++ paper/paper.md
@@ -62,8 +62,7 @@ library(gsMAMS)
 
 ```R
 set.seed(1234)
-design_cont(delta0 = 0.178,delta1 = 0.545,alpha = 0.05,beta = 0.1,K=3,
-frac = c(0.5,1))
+design_cont(delta0 = 0.178, delta1 = 0.545, alpha = 0.05, beta = 0.1, K = 3, frac = c(0.5, 1))
 ```
 
 For FWER and Stagewise FWER:
@@ -71,14 +70,12 @@ For FWER and Stagewise FWER:
 The operating characteristics of the trial can be generated using the op_power_cont and op_fwer_cont functions for power under alternative hypothesis and FWER under global null hypothesis respectively. Most of the arguments in the function are similar to size and SCPRT functions with the exception of number of simulations(nsim) and seed number(seed).
 
 ```R
-op_fwer_cont(alpha=0.05,beta=0.1,K=3,frac=c(0.5,1),delta0=0.178,
-delta1=0.545,nsim=10000,seed=10)
+op_fwer_cont(alpha = 0.05, beta = 0.1, K = 3, frac = c(0.5, 1), delta0 = 0.178, delta1 = 0.545, nsim = 10000, seed = 10)
 ``` 
 
 For Power and Stagewise Power:
 ```R
-op_power_cont(alpha=0.05,beta=0.1,K=3,frac=c(0.5,1),delta0=0.178,
-delta1=0.545,nsim=10000,seed=10)
+op_power_cont(alpha = 0.05, beta = 0.1, K = 3, frac = c(0.5, 1), delta0 = 0.178, delta1 = 0.545, nsim = 10000, seed = 10)
 ```
 
 ## Ordinal Outcome
@@ -86,8 +83,7 @@ For ordinal outcome, we will consider ASCLEPIOS trial, a phase II trial for pati
 The sample size calculation is based on a one-sided FWER of 5\% and a power of 90\%. Based on the trial characteristics, we will design the trial for a three-stage design. The design parameters for a five-arm (K = 4) trial can be calculated using design_ord function and the arguments in the function correspond to probability of outcomes in control group(prob), odds ratio of ineffective treatment group vs control(or0), odds ratio of effective treatment group vs control(or) and the remaining arguments are similar to the design_cont function for continuous outcome.
 
 ```R
-design_ord(prob=c(0.075, 0.182, 0.319, 0.243, 0.015, 0.166),or=3.06,
-or0=1.32,alpha=0.05,beta=0.1,K=4,frac = c(1/3,2/3,1))
+design_ord(prob = c(0.075, 0.182, 0.319, 0.243, 0.015, 0.166), or = 3.06, or0 = 1.32, alpha = 0.05, beta = 0.1, K = 4, frac = c(1/3, 2/3, 1))
 ```
 The operating characteristics can be generated using the functions op_fwer_ord and op_power_ord which are similar to that of continuous outcome.
 
@@ -97,21 +93,18 @@ For survival outcome, we will consider a MAMS trial with five arms (four treatme
[…]
