### 2024-03-15 Tpatni719/gsMAMS#2 @Tpatni719
Thank you for the suggestion! I have made the syntax consistent as per your recommendation.
