### 2024-03-15 Tpatni719/gsMAMS#3 @Tpatni719
The design functions of the trial are indeed deterministic, but for generating the operating characteristics of the trial, we need the seed for reproducibility.
