### 2024-03-15 Tpatni719/gsMAMS#4 @RhysPeploe
Automated Tests
https://github.com/openjournals/joss-reviews/issues/6322

From the JOSS checklist - **Automated tests: Are there automated tests or manual steps described so that the functionality of the software can be verified?**

Can some tests be created for the functions? Some simple ones that one the examples, ensuring no errors occur will be sufficient for me to check this list
