### 2024-03-15 Tpatni719/gsMAMS#3 @RhysPeploe
Ahh very well! Then I do not have concern with the seed. The results would be a good extra but that will not stop the checklist being completed - thanks
