### 2024-03-15 #6322 @RhysPeploe
Hi all,
Just to update on my checklist for Substanial Scholarly Effort - that depends on the comment above https://github.com/openjournals/joss-reviews/issues/6322#issuecomment-1997943810
@ppxasjsm thoughts on this? If you are happy with the commented code then I'll tick this off

And for Automated Tests, there does not appear to be any currently so when Issue #4  is resolved then that will be ticked off

Just those two issues, after which my checklist will be complete!
