### 2024-03-15 Tpatni719/gsMAMS#4 @Tpatni719
Thank you for the comment! So, to check the veracity of these functions, we check whether we reached the desired power or Familywise error rates in the simulations. To give an example:
```r
op_fwer_cont(alpha=0.05,beta=0.1,K=3,frac=c(0.5,1),delta0=0.178, delta1=0.545,nsim=10000,seed=10)
$FWER
[1] 0.05 (Reached the desired FWER)
$`Stagewise FWER`
look1 look2
0.0050 0.0466 (should add up to around 0.05)
```
In the above results, we have reached the desired type I error of 0.05 which makes sure that the design and op functions are working correctly. Similarly, we can use the same logic in power characteristic.
Does that answer your question?
