### 2024-03-18 Tpatni719/gsMAMS#4 @RhysPeploe
Are these simulations ran manually or automatically? It would be good to have these checks ran when you go through the CMD build stage in the package check
