### 2024-03-18 Tpatni719/gsMAMS#4 @Tpatni719
These simulations are run using the operating characteristics functions provided in the package. So, when we are designing a trial, we want to make sure that the trial parameters are such which gives the desired type I and type II error rate and in order to do that people routinely run simulations to check the operating characteristics of the trial. But sometimes, we can't reach the desired power and FWER (for example if your power is 0.9 then you might only reach 0.89 or 0.88) because of random variation or the number of simulations that you have executed. Please see the below snapshot in which our beta=0.1(Power=1-beta=0.9) but we only reached power of 0.893.
![image](https://github.com/Tpatni719/gsMAMS/assets/13623648/9a4e0b85-6792-48ee-95f7-52e4433be928)

Therefore, it is kind of difficult to create these checks but as long as we are close to desired power or FWER then we know that the trial design is working correctly and sometimes, investigators want to tweak these parameters to such an extent that it will be  difficult to reach the desired power because of very low effect size, type I and type II error rate just to get a feel of the trial design.

Our package provides functions for following two stages:
1. To design the trial using the design functions.
2. To check the veracity of design using the operating characteristics functions.
