### 2024-03-19 #6322 @RhysPeploe
@ppxasjsm would you also be able to check issue #4 - automated tests and advise whether that satifies the chechlist list, please?
