### 2024-03-25 Tpatni719/gsMAMS#5 @njtierney
JOSS Review: License in DESCRIPTION does not match license provided in LICENSE.md 
As part of the review in  https://github.com/openjournals/joss-reviews/issues/6322

The [License provided in DESCRIPTION](https://github.com/Tpatni719/gsMAMS/blob/main/DESCRIPTION#L7) is GPL-3

However [LICENSE.md](https://github.com/Tpatni719/gsMAMS/blob/main/LICENSE.md) is MIT. Which license does the software use?

{gsMAMS} imports [mvtnorm](https://cran.r-project.org/web/packages/mvtnorm/index.html) which is GPL-2 - usage of this software is [copyleft, so requiring release under GPL](https://www.tldrlegal.com/license/gnu-general-public-license-v2), so MIT does not seem appropriate if the authors want to keep using {mvtnorm}
