### 2024-03-25 #6322 @njtierney
Regarding

> Does this submission meet the scope eligibility described in the [JOSS guidelines](https://joss.readthedocs.io/en/latest/submitting.html#substantial-scholarly-effort)

I note that the repository's first commit was Jan 10 - which makes it 3 months old currently but only 1 month old when submitted in February - however based on other metrics like lines of code it seems that this is a substantial piece of software. Measuring things is always hard and I do not like to reduce software to something like lines of code and age, but since they are in the guidelines I felt it was worth mentioning.

> Whether the software is sufficiently useful that it is likely to be cited by your peer group.

I think that the paper.md requires some work to more clearly demonstrate the use case and audience for this work. I will write more substantial comments on this in another comment.
