### 2024-03-25 Tpatni719/gsMAMS#6 @njtierney
JOSS Review: Functionality: Have the functional claims of the software been confirmed?
I think that to demonstrate the functional claims of the software there should be a short form vignette which describes how this software solves a problem, and work through how you would expect a user to use this software.

See the [vignettes chapter](https://r-pkgs.org/vignettes.html) in the R packages book for more details on this.

Review link: https://github.com/openjournals/joss-reviews/issues/6322
