### 2024-03-25 Tpatni719/gsMAMS#8 @njtierney
JOSS Review: Code Style and naming arguments of functions
## Using a style helper

To help standardise the code display, I think it would be worthwhile for the authors to run code like:

```r
styler::style_pkg(".")
```

This would make the package easier to read, add necessary whitespace, and remove large chunks of whitespace throughout the package.

## Consistency in naming of functions

Functions in the package are not named consistently. For example:

- `SCPRT()`
- `Size_ord()`
- `logranktest()`
- `Size_ord()`

I suggest that the authors stick to `lower_case_snake_case()` which uses all lower case functions, and `_` to separate words. The functions above would then be rewritten as:

- `scprt()`
- `size_ord()`
- `log_rank_test()`
- `size_ord()`

## SCPRT

Why is this function called this? The title and description for the function are:

```r
#' @title  Efficacy and Futility boundary values
#' @description This function calculates the upper and lower bound values for MAMS trial.
```

If I was trying to remember this function, why would SCPRT come to mind? Perhaps the documentation could explain this?

Or does this not matter because it is an internal function?

## Consistency in naming of arguments

Arguments to the functions should be lowercase - there is a mix of upper and lowercase letters, which might be hard for the user to remember.

## Naming arguments

In addition, the package frequently does not name the arguments of functions in the examples. For example:

```r
op_fwer_surv(20,0.05,0.1,4,c(1/2,1),1,0.75,12,40,20,1,0,12)
```

In addition internally the `Size_surv()` function often does not use named arguments.

JOSS review: https://github.com/openjournals/joss-reviews/issues/6322
