### 2024-03-25 Tpatni719/gsMAMS#9 @njtierney
JOSS Review: Documentation - A statement of need & Example Usage
RE:

> A statement of need: Do the authors clearly state what problems the software is designed to solve and who the target audience is?
> Example usage: Do the authors include examples of how to use the software (ideally to solve real-world analysis problems).
>

I think the authors could address this by adding a couple of sentences that answer these questions (taken from the [README section of the R packages book](https://r-pkgs.org/other-markdown.html#sec-readme)):

> 1. Why should I use it?
> 2. How do I use it?
> 3. How do I get it?

This section then goes on to provide a structure for the README:

> A paragraph that describes the high-level purpose of the package.
>
> An example that shows how to use the package to solve a simple problem.
>
> Installation instructions, giving code that can be copied and pasted into R.
>
> An overview that describes the main components of the package. For more complex packages, this will point to vignettes for more details. This is also a good place to describe how your package fits into the ecosystem of its target domain.

Based on the current paragraph of information in the README, they state:

> R package provides a good platform for designing and planning trials with multiple treatment arms. The package provides functionality for designing trials with continuous, ordinal and survival outcomes which offers great flexibility and makes it a comprehensive toolkit to handle different kinds of trials.

To me this indicates that this is to do with medical trials? Or clinical trials, perhaps? I am not sure what "multiple treatment arms" means - it makes me wonder what "single treatment arms" are? Based on what they have said here, I would also expect to see some example workflows of how this package works with designing trials, and the sorts of questions that come up when designing trials, and how this fits in to the puzzle.

The authors provide an example of usage:

```r
#For designing a trial with continuous outcome.
design_cont(delta0 = 0.178, delta1 = 0.545, alpha = 0.05, beta = 0.1, K = 3, frac = c(0.5, 1))

#Generating operating characteristics of the trial.
op_power_cont(alpha = 0.05, beta = 0.1, K = 3, frac = c(0.5, 1), delta0 = 0.178, delta1 = 0.545, nsim = 10000, seed = 10)
```

But I think that they should first describe what kind of data continuous or operating characteristics means? Perhaps with an example to an existing study, or some existing example data.

The authors then state:

> For detailed usage of the package, please see the paper in the repository.

I think the `paper.md` file should be converted into a vignette that gives more detail about the usage of the software.

I think that by writing more substantial documentation this would also help cover the "example usage":Do the authors include examples of how to use the software (ideally to solve real-world analysis problems).

Link to review: https://github.com/openjournals/joss-reviews/issues/6322
