### 2024-03-25 Tpatni719/gsMAMS#5 @Tpatni719
Thank you for pointing out the discrepancy! I will change the license to GPL-3.
