### 2024-03-25 5300ada
Changed the license to GPL-3
