### 2024-03-25 #6322 @Tpatni719
> Regarding
>
> > Contribution and authorship: Has the submitting author (@Tpatni719) made major contributions to the software? Does the full list of paper authors seem appropriate and complete?
>
> It is not clear to me how to evaluate the input of the author co-authors on this paper, there is no "contributions" or description of what other co-workers have done, and looking at the [git contributions](https://github.com/Tpatni719/gsMAMS/graphs/contributors) it seems like there are only commits from [@Tpatni123](https://github.com/Tpatni123) and [@Tpatni719](https://github.com/Tpatni719) who I think are the same author?

The other two authors had worked on the methodology code(the manuscript mentioned in the package at CRAN) and provided contributions in making the package user-friendly.
