### 2024-03-25 #6322 @Tpatni719
> Regarding
>
> > Does this submission meet the scope eligibility described in the [JOSS guidelines](https://joss.readthedocs.io/en/latest/submitting.html#substantial-scholarly-effort)
>
> I note that the repository's first commit was Jan 10 - which makes it 3 months old currently but only 1 month old when submitted in February - however based on other metrics like lines of code it seems that this is a substantial piece of software. Measuring things is always hard and I do not like to reduce software to something like lines of code and age, but since they are in the guidelines I felt it was worth mentioning.

The package was published at CRAN on 2023-05-04 and currently, it has around 5000 downloads.
