### 2024-03-25 Tpatni719/gsMAMS#6 @Tpatni719
In the paper, we have provided comprehensive examples explaining each argument of every function so that the user can easily use the package based on these examples.
