### 2024-03-25 Tpatni719/gsMAMS#8 @Tpatni719
Thank you for the comments!
So, following are the mains functions:
![image](https://github.com/Tpatni719/gsMAMS/assets/13623648/73bf5e0d-e5f4-4c3d-91c7-bd70db55483f)

And the remaining functions are internal functions.
Regarding naming the arguments in the examples, I tried to do that but while developing the package for CRAN, the examples were getting trimmed in the package manual.
I have standardized the code display based on your recommendation. Thank you!
