+++ R/op_fwer_cont.R
@@ -1,6 +1,3 @@
- 
-
-
 #' @title Provides operating characteristics of group sequential MAMS trial for continuous outcome under null hypothesis
 #' @description Computes FWER and other characteristics for group-sequential MAMS trial for continuous outcome.
 #' @param alpha Type I error.
@@ -12,279 +9,272 @@
 #' @param nsim Number of simulations.
 #' @param seed Random seed number.
 #' @import stats
-#'
 #' @return A list of FWER, stage-wise type I error, average sample size used per arm, stopping probability, probability of futility.
-#'
 #' @examples
-#' op_fwer_cont(0.05,0.1,2,c(0.5,1),0.178,0.545,15,1)
+#' op_fwer_cont(0.05, 0.1, 2, c(0.5, 1), 0.178, 0.545, 15, 1)
 #' @export
 
-op_fwer_cont<-function(alpha,beta,K,frac,delta0,delta1,nsim,seed){
+op_fwer_cont <- function(alpha, beta, K, frac, delta0, delta1, nsim, seed) {
+  if (K <= 1) {
+    stop("K should be greater than 1.")
+  }
 
-  if (K<=1) {stop("K should be greater than 1.")}
-  
-  if(length(frac)==1){stop("The length of frac should be greater than 1.")}
-  
-  j<-length(frac)
-  
-  bound<-SCPRT(alpha = alpha,K=K,frac = frac)
+  if (length(frac) == 1) {
+    stop("The length of frac should be greater than 1.")
+  }
 
-  l<-Size_cont(delta0=delta0,delta1=delta1,alpha = alpha,beta = beta,K=K)
-  #l<-84
+  j <- length(frac)
 
-  mu0=0
-  mu1<-0
-  mu4<-0
+  bound <- SCPRT(alpha = alpha, K = K, frac = frac)
 
-  
-  sp<-numeric()
-  pf<-numeric()
-  s2<-numeric(length=j)
-  s1<-0
-  asn<-0
-  frac<-frac
-  n<-numeric(length = j)
-  for(i in 1:j) {
-    n[i]<-ceiling(l*frac[i])
+  l <- Size_cont(delta0 = delta0, delta1 = delta1, alpha = alpha, beta = beta, K = K)
+  # l<-84
+
+  mu0 <- 0
+  mu1 <- 0
+  mu4 <- 0
+
+
+  sp <- numeric()
+  pf <- numeric()
+  s2 <- numeric(length = j)
+  s1 <- 0
+  asn <- 0
+  frac <- frac
+  n <- numeric(length = j)
+  for (i in 1:j) {
+    n[i] <- ceiling(l * frac[i])
   }
 
 
-  #a<-cbind.data.frame(c(-0.630 ,-0.630 ,-0.630 ,-0.630 ),c(0.437,0.437,0.437,0.437),c(2.161,2.161,2.161,2.161))
-  #b<-cbind.data.frame(c(3.126,3.126,3.126,3.126),c(3.092,3.092,3.092,3.092),c(2.161,2.161,2.161,2.161))
+  # a<-cbind.data.frame(c(-0.630 ,-0.630 ,-0.630 ,-0.630 ),c(0.437,0.437,0.437,0.437),c(2.161,2.161,2.161,2.161))
+  # b<-cbind.data.frame(c(3.126,3.126,3.126,3.126),c(3.092,3.092,3.092,3.092),c(2.161,2.161,2.161,2.161))
 
-  a<-data.frame()
-  for(i in 1:K){
-    a<-rbind.data.frame(a,bound$lshape)
+  a <- data.frame()
+  for (i in 1:K) {
[…]
