### 2024-03-25 Tpatni719/gsMAMS#7 @Tpatni719
Thank you for the comments and sharing the name of the package!
Regarding the unit tests, I have discussed it with @RhysPeploe. Please see issue #4.
I have fixed the hidden files issue based on your recommendation.
And thank you so much for the usethis::use_package("stats") pointer. It is really helpful and I will implement this and the remaining suggestions in my future builds.
