### 2024-03-25 Tpatni719/gsMAMS#9 @Tpatni719
Thank you for the comments!
I have explained the example mentioned in the Readme section and the operating characteristics is a common jargon in clinical trial settings so I don't think it is necessary to expound it. Regarding adding the vignette, I think the paper will be a good addition as a vignette at CRAN and now, the results of all the examples along with their comprehensive explanation is also included in the paper.
