### 2024-03-25 06060fb
Added results of the examples with detailed explanation.
