+++ paper/paper.md
@@ -63,7 +63,25 @@ library(gsMAMS)
 ```R
 set.seed(1234)
 design_cont(delta0 = 0.178, delta1 = 0.545, alpha = 0.05, beta = 0.1, K = 3, frac = c(0.5, 1))
+
+## $`Sample size`
+##                                            Stage 1 Stage 2
+## Cumulative sample size for treatment group      40      79
+## Cumulative sample size for control group        40      79
+## 
+## $`Maximum total sample size for the trial`
+## [1] 316
+## 
+## $`Boundary values`
+##             Stage 1 Stage 2
+## Lower bound   0.006   2.062
+## Upper bound   2.910   2.062
+
 ```
+
+The design output shows the cumulative sample size for treatment and control groups at each stage. The SCPRT lower and upper boundaries are (0.006, 2.062) and (2.91, 2.062) respectively.
+Based on the design parameters, the first interim analysis can be conducted after the enrollment of 40 patients in the control arm. If the test statistic $Z_{k,l}<0.006$, the $k_{th}$ arm is rejected for futility at $1^{st}$ stage and the trial continues with the remaining treatment arms and the control. If the test statistic 0.006 $\le$ $Z_{k,l}$ $\le$ 2.91 for $k$=1,2,3 then the trial continues to the next stage and 39 patients are further enrolled per arm. If $Z_{k,1} >2.91$ for some $k$, the trial is terminated and the arm with maximum value of \{$Z_{k,1}$, $k$=1,2,3\} would be recommend for further study.
+
 
 For FWER and Stagewise FWER:
 
@@ -71,12 +89,55 @@ The operating characteristics of the trial can be generated using the op_power_c
 
 ```R
 op_fwer_cont(alpha = 0.05, beta = 0.1, K = 3, frac = c(0.5, 1), delta0 = 0.178, delta1 = 0.545, nsim = 10000, seed = 10)
-``` 
+
+## $FWER
+## [1] 0.05
+## 
+## $`Stagewise FWER`
+##  look1  look2 
+## 0.0050 0.0466 
+## 
+## $`Stopping probability under null`
+##  look1  look2 
+## 0.2599 0.7401 
+## 
+## $`Probability of futility under null`
+##  look1  look2 
+## 0.2549 0.6949 
+## 
+## $`Average sample size used per arm under null`
+## [1] 61.645
+
+```
+
+Based on the simulation results, the type I error at the first interim analysis is 0.5\% and at the second interim analysis, it is around 4.6\%. Therefore, the overall type I error of the trial is close to 5\%. The sample size required for the trial was 79 patients per arm but the trial used around an average of 62 subjects per arm. The stopping probability(probability of stopping the trial either due to futility or efficacy) should add up to 1 which is the case here and since this is under null configuration, the probability of futility(probability of stopping the trial when all the treatment arms becomes futile in the trial) should be around 95\% and it holds true in this case. 
+
 
 For Power and Stagewise Power:
 ```R
 op_power_cont(alpha = 0.05, beta = 0.1, K = 3, frac = c(0.5, 1), delta0 = 0.178, delta1 = 0.545, nsim = 10000, seed = 10)
+
+## $Power
+## [1] 0.893
+## 
+## $`Stagewise Power`
+##  look1  look2 
+## 0.3126 0.5804 
+## 
+## $`Stopping probability under alternative`
+##  look1  look2 
+## 0.3258 0.6742 
+## 
+## $`Probability of futility under alternative`
+##  look1  look2 
+## 0.0035 0.0821 
+## 
+## $`Average sample size used per arm under alternative`
+## [1] 62.652
+
 ```
[…]
