+++ paper/paper.md
@@ -80,7 +80,7 @@ design_cont(delta0 = 0.178, delta1 = 0.545, alpha = 0.05, beta = 0.1, K = 3, fra
 ```
 
 The design output shows the cumulative sample size for treatment and control groups at each stage. The SCPRT lower and upper boundaries are (0.006, 2.062) and (2.91, 2.062) respectively.
-Based on the design parameters, the first interim analysis can be conducted after the enrollment of 40 patients in the control arm. If the test statistic $Z_{k,l}<0.006$, the $k_{th}$ arm is rejected for futility at $1^{st}$ stage and the trial continues with the remaining treatment arms and the control. If the test statistic 0.006 $\le$ $Z_{k,l}$ $\le$ 2.91 for $k$=1,2,3 then the trial continues to the next stage and 39 patients are further enrolled per arm. If $Z_{k,1} >2.91$ for some $k$, the trial is terminated and the arm with maximum value of \{$Z_{k,1}$, $k$=1,2,3\} would be recommend for further study.
+Based on the design parameters, the first interim analysis can be conducted after the enrollment of 40 patients in the control arm. If the test statistic $Z_{k,l}<0.006$, the $k_{th}$ arm is rejected for futility at $1^{st}$ stage and the trial continues with the remaining treatment arms and the control. If the test statistic 0.006 $\le$ $Z_{k,l}$ $\le$ 2.91 for $k$=1,2,3 then the trial continues to the next stage and 39 patients are further enrolled per arm. If $Z_{k,1} >2.91$ for some $k$, the trial is terminated and the arm with maximum value of {$Z_{k,1}$, $k$=1,2,3} would be recommend for further study.
 
 
 For FWER and Stagewise FWER:
@@ -160,7 +160,7 @@ design_ord(prob = c(0.075, 0.182, 0.319, 0.243, 0.015, 0.166), or = 3.06, or0 = 
 ## Upper bound   3.126   3.092   2.161
 
 ```
-Based on the design parameters, the first interim analysis can be conducted after the enrollment of 21 patients(information time is 21/62 which is approximately 1/3 at this stage) in the control and treatment arms. If the test statistic $Z_{k,l}<-0.63$, the $k_{th}$ arm is rejected for futility at $1^{st}$ stage and the trial continues with the remaining treatment arms and the control. If the test statistic -0.63 $\le$ $Z_{k,l}$ $\le$ 3.126 for $k$=1,2,3,4 then the trial continues to the next stage and 21 patients are further enrolled per arm. If $Z_{k,1} >3.126$ for some $k$, the trial is terminated and the arm with maximum value of \{$Z_{k,1}$, $k$=1,2,3,4\} would be recommend for further study. Similar procedure is followed if the trial goes to second stage of interim analysis.
+Based on the design parameters, the first interim analysis can be conducted after the enrollment of 21 patients(information time is 21/62 which is approximately 1/3 at this stage) in the control and treatment arms. If the test statistic $Z_{k,l}<-0.63$, the $k_{th}$ arm is rejected for futility at $1^{st}$ stage and the trial continues with the remaining treatment arms and the control. If the test statistic -0.63 $\le$ $Z_{k,l}$ $\le$ 3.126 for $k$=1,2,3,4 then the trial continues to the next stage and 21 patients are further enrolled per arm. If $Z_{k,1} >3.126$ for some $k$, the trial is terminated and the arm with maximum value of {$Z_{k,1}$, $k$=1,2,3,4} would be recommend for further study. Similar procedure is followed if the trial goes to second stage of interim analysis.
 
 
[…]
