### 2024-03-25 Tpatni719/gsMAMS#9 @njtierney
Great!

I still think that this first section needs to be unpacked with more detail:

> ggMAMS R package provides a good platform for designing and planning trials with multiple treatment arms. The package provides functionality for designing trials with continuous, ordinal and survival outcomes which offers great flexibility and makes it a comprehensive toolkit to handle different kinds of trials.

Specifically, here is a small breakdown of what I mean.

> ggMAMS R package provides a good platform for designing and planning trials with multiple treatment arms.

- What does "good platform" mean?
- Designing and planning what kind of trials?
- What are multiple treatment arms? What are single treatment arms?
- What specific problem does this solve within these?
- How does it help with design and planning?
- What information does it help provide the user of the package?

> The package provides functionality for designing trials with continuous, ordinal and survival outcomes which offers great flexibility and makes it a comprehensive toolkit to handle different kinds of trials.

- Can you tell me more about why there are certain outcomes?
- Can you give me an example of the different kinds of trials? Medical trials? Drug trials? Just a couple of examples would be useful.

> I have explained the example mentioned in the Readme section and the operating characteristics is a common jargon in clinical trial settings so I don't think it is necessary to expound it.

Could you use a README.Rmd file so we can see the output of the example above? `usethis::use_readme_rmd()` will help generate the file.

> The operating characteristics of the trial can be generated using the op_power_cont() functions for power under alternative hypothesis.

Can you describe how these are used, and what they do?
