### 2024-03-25 Tpatni719/gsMAMS#4 @njtierney
Regarding tests, I would suggest reading this section of the [R packages book on testing](https://r-pkgs.org/testing-basics.html).

There are lots of ways to think about tests, but in one sense, these tests are useful in understanding what the outputs of the functions are. I know that this doesn't necessary tell you if the test did the right thing, but it is really useful when you go back to your package and make changes - your tests will identify if the output changes and you won't need to run your own manual tests in the console to understand how it might have changed.

Here's an example of how you could create tests for `design_cont()`:

```r
usethis::use_testthat()
# create the test file for design_cont
use_test("design_cont")

continuous_design_output <- design_cont(
  delta0 = 0.178,
  delta1 = 0.545,
  alpha = 0.05,
  beta = 0.1,
  K = 4,
  frac = c(1 / 2, 1)
)

# check names remains the same
expect_snapshot(
  names(continuous_design_output)
)

# check there are 4 values returned
expect_length(
  as.numeric(continuous_design_output$`Sample size`),
  4
)

# check that the values are greater than or equal to 420
expect_gte(
  as.numeric(continuous_design_output$`Maximum total sample size for the trial`),
  420
)

# expect boundary values are double
expect_type(
  continuous_design_output$`Boundary values`,
  type = "double"
  )

```
