### 2024-03-25 Tpatni719/gsMAMS#8 @njtierney
> Regarding naming the arguments in the examples, I tried to do that but while developing the package for CRAN, the examples were getting trimmed in the package manual.

I'm not sure that I understand this issue?

# Function naming

I still think it is actually really important to have consistent naming of your functions in the package, especially given that
 `SCPRT()` appears to be a commonly used function. What does SCPRT stand for? If a user wanted to use your code or develop it and understand it I think it is worthwhile to make the function names have a consistent format. I suggest that you use `lower_snake_case()` as suggested in the previous comment.

# Regarding closing issues

I also think that in the spirit of this review it would be best to wait closing the issue and confirming if the issue has been completed before closing it.
