### 2024-03-25 Tpatni719/gsMAMS#8 @njtierney
# Arguments to the functions should be lowercase

There is a mix of upper and lowercase letters, which might be hard for the user to remember I suggest making all of these lowercase
