### 2024-03-25 Tpatni719/gsMAMS#6 @njtierney
I think that the paper should be turned into a vignette as well. The paper.md will not be shipped with the package.

> In the paper, we have provided comprehensive examples explaining each argument of every function so that the user can easily use the package based on these examples.

Have you read the [vignette](https://r-pkgs.org/vignettes.html) chapter of the R Packages book? This provides examples of how to create a vignette. In addition this will provide the output of the code that you use.
