### 2024-03-25 Tpatni719/gsMAMS#9 @Tpatni123
Sure, I can add some details for the mentioned points in the Readme section(most of them are already mentioned in the paper) but I don't understand what you meant by "why there are certain outcomes".
