### 2024-03-25 Tpatni719/gsMAMS#8 @Tpatni123
I'm not sure that I understand this issue?
>>
![image](https://github.com/Tpatni719/gsMAMS/assets/42519735/f59a32f0-9220-42d3-a092-bdd84d8bf0d8)
So, if I include the names of the arguments then the function goes beyond the margins of the PDF manual.

Function naming and Arguments to the functions should be lowercase
>> I will do that but they all are internal functions and all the detailed information is provided in the description if someone wants to comprehend and use the code.
![image](https://github.com/Tpatni719/gsMAMS/assets/42519735/12a1f402-dec1-4847-867d-e15554bd3baa)

Regarding closing issues
>> My apologies for that. I will wait for the confirmation.
