### 2024-03-26 Tpatni719/gsMAMS#8 @njtierney
> So, if I include the names of the arguments then the function goes beyond the margins of the PDF manual.

So the solution to this is to use linebreaks in your code - in general we recommend not exceeding an 80 character width, as it can run past the margins on a text editor and mean that users will need to horizontally scroll to read code.

So it would change like so:

```r
#' @examples
#' design_ord(0.05, 0.1, K = 4, c(0.075, 0.182, 0.319, 0.243, 0.015, 0.166), or = 3.06, or0 = 1.32, c(1 / 2, 1))
```

```r
#' @examples
#' design_ord(alpha = 0.05,
#'            beta = 0.1,
#'            K = 4,
#'            prob = c(0.075, 0.182, 0.319, 0.243, 0.015, 0.166),
#'            or = 3.06,
#'            or0 = 1.32,
#'            frac = c(1 / 2, 1))

# or even

#' @examples
#' design_probs <- c(0.075, 0.182, 0.319, 0.243, 0.015, 0.166)
#' design_frac <- c(1 / 2, 1)
#' design_ord(alpha = 0.05,
#'            beta = 0.1,
#'            K = 4,
#'            prob = design_probs,
#'            or = 3.06,
#'            or0 = 1.32,
#'            frac = design_frac)

```

In general it is considered good practice to name your arguments. There are two reasons for this:

1. Naming arguments makes it explicit which argument is being used
2. R uses argument names first, and if they aren't specified uses position (the order).

If you don't name the arguments then you are expecting the user (and your future self) to remember what the seven numbers mean in order, and this is a very easy way to end up making potentially critical errors regarding something as important as clinical trial design.

If you have a mix of argument names and change the order around then you can end up with unpredictable output:

```r
# what does 0.1  and 0.05 correspond to?
design_ord(K = 4,
           or = 3.06,
           0.1,
           prob = c(0.075, 0.182, 0.319, 0.243, 0.015, 0.166),
           0.05,
           or0 = 1.32,
           frac = c(1 / 2, 1))

# is it the same as this?
design_probs <- c(0.075, 0.182, 0.319, 0.243, 0.015, 0.166)
design_frac <- c(1 / 2, 1)
design_ord(alpha = 0.05,
           beta = 0.1,
           K = 4,
           prob = design_probs,
           or = 3.06,
           or0 = 1.32,
           frac = design_frac)
```

# Re SCPRT

Thanks for explaining SCPRT - I would suggest making it all lower case if you can? The reason is that if you have consistent style in your package it becomes easier for others to predict and use your code without needing to remember if one particular function has mixed case, or all upper case. It's a small thing but it makes a difference for future users and possible contributors to have consistent styling in your code.
