### 2024-03-26 Tpatni719/gsMAMS#9 @njtierney
It's great that you've mentioned them in the paper! But most users will read the README first, so I think it's important to make it as easy as possible for users of a variety of backgrounds to go from the README to understand how to use the package.

> "why there are certain outcomes".

Sorry I should have been clearer - some examples of the the type of data or trial that contains continuous, survival, or ordinal outcomes.
