### 2024-03-26 Tpatni719/gsMAMS#6 @Tpatni123
Thanks for the link! I will add the vignette.
