### 2024-03-26 13a89a7
Made changes in the structure of the code
