### 2024-03-26 Tpatni719/gsMAMS#8 @Tpatni719
All the functions are now in lowercase and the names of the arguments are added in the examples.
