### 2024-03-26 Tpatni719/gsMAMS#9 @Tpatni719
The details have been added.
