### 2024-03-27 Tpatni719/gsMAMS#9 @njtierney
Thanks!

Just to confirm, it looks like you are manually updating the README.md with R code output - are you using a README.Rmd file or?
