### 2024-03-27 Tpatni719/gsMAMS#8 @njtierney
Brilliant, thank you!
