### 2024-03-27 Tpatni719/gsMAMS#9 @njtierney
That is a lot of copying and pasting! You might really benefit from using Rmarkdown files, have you seen then before?

If you use:

```r
# if not already installed
# install.packages("usethis")
usethis::use_readme_rmd()
```

This will generate a special file that you can provide the code in, then click the "knit" button in Rstudio, it will then mean you won't need to carefully copy and paste and format the code.
