+++ paper/paper.md
@@ -28,16 +28,17 @@ bibliography: paper.bib
 ---
 
 # Summary
-A multi-arm trial allows simultaneous comparison of multiple experimental treatments with a common control and provides a substantial efficiency advantage compared to conventional randomized controlled trial. A multi-stage trial allows multiple interim looks at the trial outcome so that the ineffective arm can be stopped early. In the current R ecosystem, `MAMS` @Jaki2019 is one of the few packages that can handle multiple stages, multiple treatment arms and different types of endpoints but the computational effort of obtaining sample size and sequential stopping boundaries is very high when the number of stages exceeds 3. More importantly, their method may lead to under powered study when the endpoint is time to event data. Therefore, we introduce the R package `gsMAMS` for designing group sequential multi-arm multi-satge (MAMS) trials with continuous, ordinal and survival outcomes which is computationally very efficient even for number of stages greater than 3. We also discuss the applications of the package.          
-
+A multi-arm trial allows simultaneous comparison of multiple experimental treatments with a common control and provides a substantial efficiency advantage compared to conventional randomized controlled trial. A multi-stage trial allows multiple interim looks at the trial outcome so that the ineffective arm can be stopped early. In the current R ecosystem, `MAMS` @Jaki2019 is one of the few packages that can handle multiple stages, multiple treatment arms and different types of endpoints but the computational effort of obtaining sample size and sequential stopping boundaries is very high when the number of stages exceeds 3. More importantly, their method may lead to under powered study when the endpoint is time to event data. Therefore, we introduce the R package `gsMAMS` for designing group sequential multi-arm multi-satge (MAMS) trials with continuous, ordinal and survival outcomes which is computationally very efficient even for number of stages greater than 3. We also discuss the applications of the package. The `gsMAMS` software facilitates the design and planning of multi-arm clinical trials, which simultaneously compare multiple experimental treatments with a common control group. It is intended to be used by clinical researchers and statisticians, so that they can determine appropriate sample sizes and stopping criteria for these trials.
 
 
 # Statement of Need
 Traditional two-arm randomized control trials are not an optimal choice when multiple experimental arms are available for testing efficacy. In such situations, a multiple-arm trial should be preferred, which allows simultaneous comparison of multiple experimental arms with a common control and provides substantial efficiency advantage. In multi-arm trials, several arms are monitored in a group sequential fashion, with ineffective arms being dropped out of the study. 
 
-Some packages that are available in R have limitations either in the number of treatment arms that can be incorporated in the package, the number of interim analyses that can be implemented in the package, or the different kinds of outcomes that the package can handle, but the `MAMS` package works well both for multiple treatment arms and multiple stages. It also works for continuous, ordinal, and survival outcomes. But the computational effort of obtaining stopping boundaries is very high when the number of stages exceeds 3. This is the major hurdle of using `MAMS` package.
+Some packages that are available in R(e.g. `adaptTest`@Vandemeulebroecke2022, `asd`@Hack2022 and `AGSDEST`@Parsons2016) have limitations either in the number of treatment arms that can be incorporated in the package, the number of interim analyses that can be implemented in the package, or the different kinds of outcomes that the package can handle, but the `MAMS` package works well both for multiple treatment arms and multiple stages. It also works for continuous, ordinal, and survival outcomes. But the computational effort of obtaining stopping boundaries is very high when the number of stages exceeds 3. This is the major hurdle of using `MAMS` package.
 
[…]
