### 2024-03-27 Tpatni719/gsMAMS#9 @Tpatni123
Thank you for sharing this but it's just that the results of operating characteristics for 10,000 simulations take time and I already had the results so that's why I just copied the results.
