### 2024-03-27 c350bfb
* use lowercase not uppercase letters throughout * tidy up examples so they are not truncated in PDF output
