+++ R/size_surv.R
@@ -22,30 +22,25 @@
 
 
 size_surv <- function(m0, alpha, beta, k, hr0, hr1, ta, tf, kappa, eta, frac) {
-  K<-k
-  HR0<-hr0
-  HR1<-hr1
-  if (K < 2 | K > 5) {
-    stop("K should be between 1 and 6.")
+  if (k < 2 | k > 5) {
+    stop("k should be between 1 and 6.")
   }
-  c <- scprt(alpha = alpha, k = K, frac = frac)$critical.value
-  hr0 <- HR0 # exp(-delta0)
-  hr1 <- HR1 # exp(-delta1)
+  c <- scprt(alpha = alpha, k = k, frac = frac)$critical.value
   lambda0 <- log(2) / m0^kappa
 
-  if (K == 2) {
+  if (k == 2) {
     lambda1 <- lambda0 * hr1
     lambda2 <- lambda0 * hr0
-  } else if (K == 3) {
+  } else if (k == 3) {
     lambda1 <- lambda0 * hr1
     lambda2 <- lambda0 * hr0
     lambda3 <- lambda0 * hr0
-  } else if (K == 4) {
+  } else if (k == 4) {
     lambda1 <- lambda0 * hr1
     lambda2 <- lambda0 * hr0
     lambda3 <- lambda0 * hr0
     lambda4 <- lambda0 * hr0
-  } else if (K == 5) {
+  } else if (k == 5) {
     lambda1 <- lambda0 * hr1
     lambda2 <- lambda0 * hr0
     lambda3 <- lambda0 * hr0
@@ -65,14 +60,14 @@ size_surv <- function(m0, alpha, beta, k, hr0, hr1, ta, tf, kappa, eta, frac) {
   }
 
 
-  if (K == 1) {
+  if (k == 1) {
     S1 <- function(t) {
       exp(-lambda1 * t^kappa)
     }
     h1 <- function(t) {
       kappa * lambda1 * t(kappa - 1)
     }
-  } else if (K == 2) {
+  } else if (k == 2) {
     S1 <- function(t) {
       exp(-lambda1 * t^kappa)
     }
@@ -85,7 +80,7 @@ size_surv <- function(m0, alpha, beta, k, hr0, hr1, ta, tf, kappa, eta, frac) {
     h2 <- function(t) {
       kappa * lambda2 * t^(kappa - 1)
     }
-  } else if (K == 3) {
+  } else if (k == 3) {
     S1 <- function(t) {
       exp(-lambda1 * t^kappa)
     }
@@ -104,7 +99,7 @@ size_surv <- function(m0, alpha, beta, k, hr0, hr1, ta, tf, kappa, eta, frac) {
     h3 <- function(t) {
       kappa * lambda3 * t^(kappa - 1)
     }
-  } else if (K == 4) {
+  } else if (k == 4) {
     S1 <- function(t) {
       exp(-lambda1 * t^kappa)
     }
@@ -129,7 +124,7 @@ size_surv <- function(m0, alpha, beta, k, hr0, hr1, ta, tf, kappa, eta, frac) {
     h4 <- function(t) {
       kappa * lambda4 * t^(kappa - 1)
     }
-  } else if (K == 5) {
+  } else if (k == 5) {
     S1 <- function(t) {
       exp(-lambda1 * t^kappa)
     }
@@ -171,27 +166,27 @@ size_surv <- function(m0, alpha, beta, k, hr0, hr1, ta, tf, kappa, eta, frac) {
     exp(-eta * t)
   }
 
-  omega0 <- 1 / (K + 1)
[…]
