### 2024-03-27 Tpatni719/gsMAMS#9 @njtierney
Sure I can appreciate that you don't want it to take too long. However you've already done the hard work and made it so users can control the number of simulations.

I've written a pull request which implements this here: https://github.com/Tpatni719/gsMAMS/pull/15

In doing so I also found a few bugs from the lowercase conversion, and took the liberty to fix those, I hope you don't mind! Those changes are found here: https://github.com/Tpatni719/gsMAMS/pull/14

So you'll need to merge https://github.com/Tpatni719/gsMAMS/pull/14 first, then https://github.com/Tpatni719/gsMAMS/pull/15.

Another bug that was caught was that in `op_power_cont` the argument `K` was changed to `p`.
