### 2024-03-27 Tpatni719/gsMAMS#9 @Tpatni123
Thank you so much for your help!
The reason you found those weird syntax because some code was kind of complicated down the pipeline so just to be cautious, I did that untidy trick so that all the code remains intact and the code won't break down after changing the argument to lower k. So, that's why I just used K=k to make sure that the code doesn't break down.
Regarding the k changed to p in operating characteristics functions, it was intentional as I used k for some other operation inside the function.
So, please let me know your thoughts.

PS: I have merged the branches and I will check the functions. Thanks for the help again!
