### 2024-03-27 17bbd50
Added example code for computational efficiency.
