+++ README.Rmd
@@ -36,6 +36,7 @@ The stable version of **gsMAMS**, v0.7.1, is available on CRAN:
 
 ```{r}
 # install.packages("gsMAMS")
+set.seed(1234)
 library(gsMAMS)
 ```
 
@@ -71,7 +72,7 @@ op_power_cont(alpha = 0.05,
               frac = c(0.5, 1), 
               delta0 = 0.178, 
               delta1 = 0.545, 
-              nsim = 100, 
+              nsim = 1000, 
               seed = 10)
 ```
 
