### 2024-03-27 9b46a66
Made the simulation results consistent with the code.
