+++ README.Rmd
@@ -72,7 +72,7 @@ op_power_cont(alpha = 0.05,
               frac = c(0.5, 1), 
               delta0 = 0.178, 
               delta1 = 0.545, 
-              nsim = 1000, 
+              nsim = 10000, 
               seed = 10)
 ```
 
