### 2024-03-28 Tpatni719/gsMAMS#4 @Tpatni123
@njtierney I have added the tests for all the functions and they check the fundamental things regarding the design and operating characteristics. But to emphasize that when we design a trial, we "always" run the operating characteristics function to check the FWER and power to make sure that they are reaching the desired level which ensures that our design parameters are correct. So, even though it is kind of a manual testing, that is the inherent nature of trial designing and planning and is routinely performed.
