### 2024-03-28 Tpatni719/gsMAMS#4 @Tpatni123
> Are these simulations ran manually or automatically? It would be good to have these checks ran when you go through the CMD build stage in the package check

@RhysPeploe I have added the unit tests with the help of @njtierney. So, please let me know if I can close this issue.
