### 2024-03-29 Tpatni719/gsMAMS#9 @njtierney
> The reason you found those weird syntax because some code was kind of complicated down the pipeline so just to be cautious, I did that untidy trick so that all the code remains intact and the code won't break down after changing the argument to lower k. So, that's why I just used K=k to make sure that the code doesn't break down.
Regarding the k changed to p in operating characteristics functions, it was intentional as I used k for some other operation inside the function. So, please let me know your thoughts.

Yes I could see this - it was one way to solve the problem, however in doing so it introduced some fragility into the package. To help resolve this I did a find and replace - which you can do with Cmd/Ctrl+F and searched for the new old arguments, being sure to update them. I hope that makes sense.

Now that there are tests, and that the README is an Rmarkdown document, you should be able to catch these errors.
