### 2024-03-29 Tpatni719/gsMAMS#6 @Tpatni719
I have added the vignette.
