+++ vignettes/gsMAMS.Rmd
@@ -3,7 +3,7 @@ title: "gsMAMS"
 author: "Tushar Patni"
 output: rmarkdown::html_vignette
 vignette: >
-  %\VignetteIndexEntry{my-vignette}
+  %\VignetteIndexEntry{gsMAMS}
   %\VignetteEngine{knitr::rmarkdown}
   %\VignetteEncoding{UTF-8}
 ---
