+++ paper/paper.bib
@@ -10,3 +10,30 @@
 }
 
 
+
+@Article{Parsons2016,
+  title = {asd: Simulations for Adaptive Seamless   Designs},
+  author = {Nick Parsons},
+  year = {2016},
+  note = {R package version 2.2},
+  url = {https://CRAN.R-project.org/package=asd},
+}
+
+@Article{Hack2022,
+  title = {AGSDest: Estimation in Adaptive Group Sequential Trials},
+  author = {Niklas Hack and Werner Brannath and Matthias Brueckner},
+  year = {2022},
+  note = {R package version 2.3.4},
+  url = {https://CRAN.R-project.org/package=AGSDest},
+}
+
+
+@Article{Vandemeulebroecke2022,
+  title = {adaptTest: Adaptive Two-Stage Tests},
+  author = {Marc Vandemeulebroecke},
+  year = {2022},
+  note = {R package version 1.1},
+  url = {https://CRAN.R-project.org/package=adaptTest},
+}
+
+
