### 2024-04-08 Tpatni719/gsMAMS#9 @Tpatni123
Hi @njtierney, did you get the chance to review the responses?
