### 2024-04-12 Tpatni719/gsMAMS#7 @njtierney
Hi @Tpatni719

Thanks for addressing some of these concerns!

Running `goodpractice::gp()` I get:

```
── GP gsMAMS ───────────────────────────────────────────────────────────────────────────

It is good practice to

  ✖ write unit tests for all functions, and all package code in
    general. 71% of code lines are covered by test cases.

    R/logranktest.R:15:NA
    R/logranktest.R:17:NA
    R/logranktest.R:18:NA
    R/logranktest.R:19:NA
    R/logranktest.R:20:NA
    ... and 483 more lines

  ✖ write short and simple functions. These functions have high
    cyclomatic complexity (>50): op_power_ord (91), op_power_cont (86),
    op_fwer_ord (82), op_fwer_cont (77), op_power_surv (73), op_fwer_surv
    (64), size_surv (54). You can make them easier to reason about by
    encapsulating distinct steps of your function into subfunctions.
  ✖ add a "URL" field to DESCRIPTION. It helps users find information
    about your package online. If your package does not have a homepage, add
    an URL to GitHub, or the CRAN package package page.
  ✖ add a "BugReports" field to DESCRIPTION, and point it to a bug
    tracker. Many online code hosting services provide bug trackers for free,
    https://github.com, https://gitlab.com, etc.
  ✖ avoid long code lines, it is bad for readability. Also, many
    people prefer editor windows that are about 80 characters wide. Try make
    your lines shorter than 80 characters

    R/design_cont.R:2:81
    R/design_cont.R:9:81
    R/design_cont.R:11:81
    R/design_cont.R:17:81
    R/design_cont.R:19:81
    ... and 228 more lines

  ✖ avoid 1:length(...), 1:nrow(...), 1:ncol(...), 1:NROW(...) and
    1:NCOL(...) expressions. They are error prone and result 1:0 if the
    expression on the right hand side is zero. Use seq_len() or seq_along()
    instead.

    R/design_cont.R:20:35
    R/design_cont.R:26:36
    R/design_ord.R:26:35
    R/design_ord.R:31:36
    R/design_surv.R:35:35
    ... and 17 more lines

  ✖ not import packages as a whole, as this can cause name clashes
    between the imported packages. Instead, import only the specific functions
    you need.
  ✖ avoid 'T' and 'F', as they are just variables which are set to the
    logicals 'TRUE' and 'FALSE' by default, but are not reserved words and
    hence can be overwritten by the user.  Hence, one should always use 'TRUE'
    and 'FALSE' for the logicals.

    R/op_fwer_ord.R:NA:NA
    R/op_fwer_ord.R:NA:NA
    R/op_fwer_ord.R:NA:NA
    R/op_fwer_ord.R:NA:NA
    R/op_power_ord.R:NA:NA
    ... and 3 more lines

────────────────────────────────────────────────────────────────────────────────────────

```

If you can address the following I can tick this off from review:

```
  ✖ add a "URL" field to DESCRIPTION. It helps users find information
    about your package online. If your package does not have a homepage, add
    an URL to GitHub, or the CRAN package package page.
  ✖ add a "BugReports" field to DESCRIPTION, and point it to a bug
    tracker. Many online code hosting services provide bug trackers for free,
    https://github.com, https://gitlab.com, etc.
  ✖ avoid 1:length(...), 1:nrow(...), 1:ncol(...), 1:NROW(...) and
    1:NCOL(...) expressions. They are error prone and result 1:0 if the
    expression on the right hand side is zero. Use seq_len() or seq_along()
    instead.

  ✖ not import packages as a whole, as this can cause name clashes
    between the imported packages. Instead, import only the specific functions
    you need.
  ✖ avoid 'T' and 'F', as they are just variables which are set to the
    logicals 'TRUE' and 'FALSE' by default, but are not reserved words and
    hence can be overwritten by the user.  Hence, one should always use 'TRUE'
    and 'FALSE' for the logicals.
```
