### 2024-04-12 Tpatni719/gsMAMS#6 @njtierney
Hi there,

For the vignette, you could make this run faster by changing the number of simulations to be lower, and then reference this in the code that users should run a certain number of simulations.
