### 2024-04-12 Tpatni719/gsMAMS#6 @njtierney
For reference it took about 5 minutes on my M1 laptop to render the vignette.
