### 2024-04-12 Tpatni719/gsMAMS#6 @Tpatni123
Thanks for letting me know! I think 5 mins is fine and I am hesitant to reduce the number of simulations as it will change all the results and as a result, the type I and type II errors will not reach the desired level. So, I think the current number of simulations is fine.
