### 2024-04-12 Tpatni719/gsMAMS#17 @njtierney
JOSS Review: Paper- Quality of writing
The JOSS Guidelines state:

> Quality of writing: Is the paper well written (i.e., it does not require editing for structure, language, or writing quality)

There are several writing improvements that I believe should be required of a journal publication for this paper. I have tried to list them for the key components of the paper but have not comprehensively listed every single change that I think is required. For instance I have mentioned there are standard ways to list functions and arguments using back ticks
 or quotes, but I have not listed every single change that is required, but hopefully it should be clear.

Overall I think that this paper requires improvement in terms of language, and also structure, and reproducibility. I will give a few examples of these improvements in this section.

# Language and structure in the paper

The paper has a few spelling mistakes - for example: "chararcteristics", "Compuational", "Desinging".

The paper needs editing for structure and writing quality. For example:

> Based on the simulation results, the probability of success/power at the first stage is 32.7\% and at the second stage is around 58.63\%. Therefore, the overall power is around 90\%.

Using `\` inside a sentence is not something I would expect in a standard of writing in a journal. I suggest replacing instances of this with the work "or". E.g., "Based on the simulation results, the probability of success (or power)..."

Additionally, using words like "around" here is informal, and I would replace instances of "around" with "approximately", or remove the word entirely.

> Traditional two-arm randomized control trials are not an optimal choice when multiple experimental arms are available for testing efficacy.

The authors should provide a citation for this statement, and should state why they are inefficient.

> Some packages that are available in R have limitations either in the number of treatment arms that can be incorporated in the package, the number of interim analyses that can be implemented in the package, or the different kinds of outcomes that the package can handle, but the `MAMS` package works well both for multiple treatment arms and multiple stages.

This is a very long sentence.

> But the computational effort of obtaining stopping boundaries is very high when the number of stages exceeds 3.

The authors should state what the computation effort here means. What is high? Is it 5 minutes, 10 minutes? A day?  Why is this a problem? Is this the only problem that this package, `gsMAMS` is solving?

> This is the major hurdle of using `MAMS` package.

I would suggest stating "The long computational time is a major drawback of using the `MAMS` package.", rather than stating "hurdle".

> The computational complexity of this package is very low.

I would argue that this package has reasonably high computational complexity, as there are many large functions in the package. What do the authors mean by the complexity being very low, and why is this relevant and important?

> The FWER is controlled by Dunnett correction, which entails finding the root of an integral of a multivariate normal distribution.

I suggest describing all acronyms in their first use. What does FWER stand for? What is a Dunnett correction, and why is that important? Could the authors provide a citation here?

> The multivariate normal densities are evaluated using the package `mvtnorm`

The `mvtnorm` package should be cited here. But more importantly, why are these densities being evaluated in the first place? The authors have provided this sentence with no context for the reasons of this being used.

> The package is efficient for any number of treatment arms and stages, but it has a limitation that it is only configured for 10 stages.

This sentence suggests that there are no limitations, but then states that there are indeed limitations. I suggest clearly stating that it can be only used for up to 10 stages? Is that what the authors mean, and does this mean that there can be 100 arms, and 10 stages?

> But for `gsMAMS` package with same trial configuration, the computational time to obtain stopping boundaries and sample size for three stages and four stages design is around 0.06 seconds for both the cases.

This sentence should be rewritten for expression - e.g.,

> However using the same trial configuration as in `MAMS`, the `gsMAMS` package takes approximately 0.06 seconds to identify stopping boundaries and sample size for both three, and four stage designs. This is approximately 7000 - 270000 times faster.

This result should then be references with a table of computation time. I would suggest using the `microbenchmark` or `bench` R packages to demonstrate the comparison.

> The design parameters of the trial can be calculated using the design_cont function and the arguments in the function correspond to standardized effect size in ineffective arm(delta0} and effective arm(delta1), type I error(alpha), type II error(beta), total number of treatment arms(K) and the information time (0.5, 1) is denoted by frac argument in the function.

When mentioning functions in the text they should go in backticks and parentheses added. e.g., `deign_cont()`. Additionally there is an errant `}` after `argm(delta0}`. Function arguments should be referenced in either backticks or quotes, e.g., `frac` or "frac". Also, there should be a space between the word and the parentheses.

All functions in the paper should be styled to have spaced around the `=` and all functions should be written with new lines for each argument, as specified in the documentation, e.g., https://github.com/Tpatni719/gsMAMS/blob/main/R/design_ord.R#L12-L18. This is important because it makes the function easier to read for the user.

> For FWER and Stagewise FWER:
>
> The operating characteristics of the trial can be generated using the op_power_cont and op_fwer_cont functions for power under alternative hypothesis and FWER under global null hypothesis respectively. Most of the arguments in the function are similar to size and SCPRT functions with the exception of number of simulations(nsim) and seed number(seed).

I'm not sure why "For FWER and Stagewise FWER:" has a column at the end, this should either be a new heading or form some part of the introduction sentence for this paragraph. Functions should also be wrapped in backticks and have parentheses added as mentioned above. The SCPRT function is also mentioned here in all capitals but no example is given and no definition is given of what this function is or means in this context.

> For ordinal outcome, we will consider ASCLEPIOS trial, a phase II trial for patients with stroke

This trial should be cited.

> We will consider the treatment worthwhile if the odds ratio between the effective and control arms is 3.06 and we set the null odds ratio to be 1.32 which is the odds ratio between the ineffective and control arms

Why is an odds ration of 3.06 considered successful? Can the authors provide some citation for these numbers?

>  The design parameters for a five-arm (K = 4) trial

The parameters I believe are now all lower case. Also the documentation states that `k` is the "Number of treatment arms." So should this be `k = 5`?

> outcomes in control group(), odds ratio of ineffective treatment group vs control(), odds ratio of effective treatment group vs control()

Why is "group()", and "control()" written like this?

> For survival outcome, we will consider a MAMS trial with five arms (four treatment arms and a control arm, K=4) and two interim looks with balanced information time (0.5, 1). The null hazards ratio is 1 and the alternative hazards ratio is 0.65.

Could the authors indicate which of these parameters link back to the function arguments?

# Reproducibility

I know that the author is concerned about the paper taking a long time to run when running so many simulations, but the author could solve this problem by specifying a small number of simulations, then writing the code to get the syntax right for the markdown, and then update the number of simulations and knit the document once.

The paper should use rmarkdown for formatting, to ensure reproducibility, as mentioned in #16. It might seem pedantic to insist that you write the results here like this instead of copying them, but in trying to replicate this paper I found that I was getting different results to what you had specified. Which makes me concerned that the results perhaps are inaccurate.

In addition, in trying to run the results of this paper I encountered several errors from the author not updating the syntax in the paper to use the latest syntax from the changes made in the software.

The numbers in the text are specified by hand based on the above results. These results should be inserted using inline R syntax to ensure the right numbers are specified. Indeed, trying to write the syntax below to ensure the results were the same as what the author had written led to me being uncertain about where certain parts of these results were referred. For example, the authors state:

> The overall stopping probability should be around 1 which is the case here

However the stopping probability is listed as:

```
$`Stopping probability under alternative`
 look1  look2
0.3334 0.6666
```

Which is not 1.

If you use rmarkdown inline syntax, you could write:

```r
power_survey_results <- op_power_surv(m0 = 20, alpha = 0.05, beta = 0.1, p = 4, frac = c(1/2, 1), HR0 = 1, HR = 0.6703, nsim = 10000, ta = 40, tf = 20, kappa = 1, eta = 0, seed = 12)
power_pct <- percent(power_survey_results$Power, accuracy = 0.01)
pr_succes_first <- percent(power_survey_results$`Stopping probability under alternative`[1], accuracy = 0.01)
pr_succes_second <- percent(power_survey_results$`Stopping probability under alternative`[2], accuracy = 0.01)

power_pct <- percent(power_survey_results$Power, accuracy = 0.01)

```

```markdown
Based on the simulation results, the probability of success/power at the first stage is `r pr_succes_first` and at the second stage is around `r pr_succes_second`. Therefore, the overall power is around `r power_pct`.
```

It also feels strange to in one instance mention the percentage down to 1 decimal place, but then for the power, to state, "the overall power is around 90%". I suggest stating 91.3%
