### 2024-04-12 #6322 @njtierney
Hi @ppxasjsm I've actually got a few concerns about reproducibility and writing quality of the paper, which I have outlined in https://github.com/Tpatni719/gsMAMS/issues/17 - just wanted to flag that here.
