### 2024-04-12 Tpatni719/gsMAMS#17 @Tpatni123
The paper should use rmarkdown for formatting, to ensure reproducibility, as mentioned in https://github.com/Tpatni719/gsMAMS/issues/16. It might seem pedantic to insist that you write the results here like this instead of copying them, but in trying to replicate this paper I found that I was getting different results to what you had specified. Which makes me concerned that the results perhaps are inaccurate.
>>> The numbers are correct. I think you didn't use the seed mentioned in the paper. I have just replicated the results for the survival outcome(power configuration)
![image](https://github.com/Tpatni719/gsMAMS/assets/42519735/025fc53b-7d78-4a96-a27a-e0035e0fd92d)

In addition, in trying to run the results of this paper I encountered several errors from the author not updating the syntax in the paper to use the latest syntax from the changes made in the software.
>>> I will update the hazard ratios in operating characteristics functions.

Indeed, trying to write the syntax below to ensure the results were the same as what the author had written led to me being uncertain about where certain parts of these results were referred. For example, the authors state:

The overall stopping probability should be around 1 which is the case here

However the stopping probability is listed as:

```
$`Stopping probability under alternative`
 look1  look2
0.3334 0.6666 `
```
Which is not 1.

>>> These two(0.3334+0.6666 ) add up  to 1. I don't know what you mean by "Which is not 1".
