### 2024-04-12 Tpatni719/gsMAMS#17 @Tpatni123
It also feels strange to in one instance mention the percentage down to 1 decimal place, but then for the power, to state, "the overall power is around 90%". I suggest stating 91.3%
>>> For a clinician who is running a trial, this is pretty much self-explanatory because the main purpose of running this operating characteristics function is to see whether we reach the desired power or not and in this case we did. So, that's why I didn't mention
the number exactly and I just mentioned that we reached the desired power which is 90%.
