+++ DESCRIPTION
@@ -14,3 +14,5 @@ Suggests:
     testthat (>= 3.0.0)
 Config/testthat/edition: 3
 VignetteBuilder: knitr
+URL: https://github.com/Tpatni719/gsMAMS
+BugReports: https://github.com/Tpatni719/gsMAMS/issues
