### 2024-04-12 Tpatni719/gsMAMS#7 @Tpatni719
I have addressed all the issues except the third one because some of the loops are not that straightforward and don't have typical start and stop points.
