+++ R/size_surv.R
@@ -14,8 +14,6 @@
 #' @return A numeric value indicating the sample size per arm.
 #' @examples
 #' size_surv(m0 = 20, hr0 = 1, hr1 = 0.65, ta = 20, tf = 40, alpha = 0.05, beta = 0.1, k = 3, kappa = 1, eta = 0, frac = c(1 / 2, 1))
-#' @import stats
-#' @import mvtnorm
 #' @keywords internal
 #' @noRd
 
@@ -428,86 +426,86 @@ size_surv <- function(m0, alpha, beta, k, hr0, hr1, ta, tf, kappa, eta, frac) {
 
 
   if (k == 2) {
-    P <- (integrate(g0, 0, tau)$value + integrate(g1, 0, tau)$value + integrate(g2, 0, tau)$value) / (k + 1)
+    P <- (stats::integrate(g0, 0, tau)$value + stats::integrate(g1, 0, tau)$value + stats::integrate(g2, 0, tau)$value) / (k + 1)
   } else if (k == 3) {
-    P <- (integrate(g0, 0, tau)$value + integrate(g1, 0, tau)$value + integrate(g2, 0, tau)$value +
-      integrate(g3, 0, tau)$value) / (k + 1)
+    P <- (stats::integrate(g0, 0, tau)$value + stats::integrate(g1, 0, tau)$value + stats::integrate(g2, 0, tau)$value +
+            stats::integrate(g3, 0, tau)$value) / (k + 1)
   } else if (k == 4) {
-    P <- (integrate(g0, 0, tau)$value + integrate(g1, 0, tau)$value + integrate(g2, 0, tau)$value +
-      integrate(g3, 0, tau)$value + integrate(g4, 0, tau)$value) / (k + 1)
+    P <- (stats::integrate(g0, 0, tau)$value + stats::integrate(g1, 0, tau)$value + stats::integrate(g2, 0, tau)$value +
+            stats::integrate(g3, 0, tau)$value + stats::integrate(g4, 0, tau)$value) / (k + 1)
   } else if (k == 5) {
-    P <- (integrate(g0, 0, tau)$value + integrate(g1, 0, tau)$value + integrate(g2, 0, tau)$value +
-      integrate(g3, 0, tau)$value + integrate(g4, 0, tau)$value + integrate(g5, 0, tau)$value) / (k + 1)
+    P <- (stats::integrate(g0, 0, tau)$value + stats::integrate(g1, 0, tau)$value + stats::integrate(g2, 0, tau)$value +
+            stats::integrate(g3, 0, tau)$value + stats::integrate(g4, 0, tau)$value + stats::integrate(g5, 0, tau)$value) / (k + 1)
   }
 
 
   if (k == 2) {
-    mu1 <- omega1 * omega0 * integrate(f1, 0, tau)$value
-    mu2 <- omega2 * omega0 * integrate(f2, 0, tau)$value
+    mu1 <- omega1 * omega0 * stats::integrate(f1, 0, tau)$value
+    mu2 <- omega2 * omega0 * stats::integrate(f2, 0, tau)$value
   } else if (k == 3) {
-    mu1 <- omega1 * omega0 * integrate(f1, 0, tau)$value
-    mu2 <- omega2 * omega0 * integrate(f2, 0, tau)$value
-    mu3 <- omega3 * omega0 * integrate(f3, 0, tau)$value
+    mu1 <- omega1 * omega0 * stats::integrate(f1, 0, tau)$value
+    mu2 <- omega2 * omega0 * stats::integrate(f2, 0, tau)$value
+    mu3 <- omega3 * omega0 * stats::integrate(f3, 0, tau)$value
[…]
