+++ R/scprt.R
@@ -6,8 +6,6 @@
 #' @return A list of three elements: critical.value, lshape(Futility), and ushape(Efficacy).
 #' @examples
 #' scprt(alpha = 0.05, k = 3, frac = c(1 / 3, 2 / 3, 1))
-#' @import stats
-#' @import mvtnorm
 #' @keywords internal
 #' @noRd
 
@@ -22,7 +20,7 @@ scprt <- function(alpha, k, frac) {
     int <- mvtnorm::pmvnorm(lower = rep(-Inf, k), upper = rep(c, k), mean = rep(0, k), sigma = Sigma)[1]
     alpha - (1 - int)
   }
-  c <- round(uniroot(root, lower = 0, upper = 999)$root, 3)
+  c <- round(stats::uniroot(root, lower = 0, upper = 999)$root, 3)
   a <- c(1.645, 2.109, 2.645, 2.953, 3.166, 3.327, 3.456, 3.562, 3.652, 3.729)
   a <- a[J]
   l <- round((c * frac - sqrt(2 * a * frac * (1 - frac))) / sqrt(frac), 3)
