+++ paper/paper.md
@@ -32,18 +32,18 @@ A multi-arm trial allows simultaneous comparison of multiple experimental treatm
 
 
 # Statement of Need
-Traditional two-arm randomized control trials are not an optimal choice when multiple experimental arms are available for testing efficacy. In such situations, a multiple-arm trial should be preferred, which allows simultaneous comparison of multiple experimental arms with a common control and provides substantial efficiency advantage. In multi-arm trials, several arms are monitored in a group sequential fashion, with ineffective arms being dropped out of the study. 
+Traditional two-arm randomized control trials are not an optimal choice when multiple experimental arms are available for testing efficacy(@PARMAR2014283). In such situations, a multiple-arm trial should be preferred, which allows simultaneous comparison of multiple experimental arms with a common control and provides substantial efficiency advantage. In multi-arm trials, several arms are monitored in a group sequential fashion, with ineffective arms being dropped out of the study. 
 
-Some packages that are available in R(e.g. `adaptTest`@Vandemeulebroecke2022, `asd`@Hack2022 and `AGSDEST`@Parsons2016) have limitations either in the number of treatment arms that can be incorporated in the package, the number of interim analyses that can be implemented in the package, or the different kinds of outcomes that the package can handle, but the `MAMS` package works well both for multiple treatment arms and multiple stages. It also works for continuous, ordinal, and survival outcomes. But the computational effort of obtaining stopping boundaries is very high when the number of stages exceeds 3. This is the major hurdle of using `MAMS` package.
+Some packages that are available in R(e.g. `adaptTest`@Vandemeulebroecke2022, `asd`@Hack2022 and `AGSDEST`@Parsons2016) have limitations either in the number of treatment arms that can be incorporated in the package, the number of interim analyses that can be implemented in the package, or the different kinds of outcomes that the package can handle, but the `MAMS` package works well both for multiple treatment arms and multiple stages. It also works for continuous, ordinal, and survival outcomes. But the computational effort of obtaining stopping boundaries is very high when the number of stages exceeds 3. The long computational time is the major drawback of using the `MAMS` package.
 
 This paper introduces the R package `gsMAMS` available at https://cran.r-project.org/web/packages/gsMAMS/ which provides functions to obtain sample size, and efficacy and futility boundaries for multiple stages and multiple experimental arms. It also provides functions that generate operating characteristics for designing the trials of continuous, ordinal, and survival outcomes. It is computationally very fast compared to the `MAMS` package, even for number of stages greater than 3.
 
-This package will serve well for clinicians, researchers and statisticians  who are involved in designing the phase-II clinical trials with multiple treatment arms because it is easy to use, provides comprehensive information regarding the trial chararcteristics and provides a substantial efficiency advantage compared to the traditional randomized controlled trial.          
+This package will serve well for clinicians, researchers and statisticians  who are involved in designing the phase-II clinical trials with multiple treatment arms because it is easy to use, provides comprehensive information regarding the trial characteristics and provides a substantial efficiency advantage compared to the traditional randomized controlled trial.          
 
 # Compuational Aspects
[…]
