+++ paper/paper.bib
@@ -36,4 +36,25 @@
   url = {https://CRAN.R-project.org/package=adaptTest},
 }
 
+@Article{PARMAR2014283,
+title = {More multiarm randomised trials of superiority are needed},
+journal = {The Lancet},
+volume = {384},
+number = {9940},
+pages = {283-284},
+year = {2014},
+issn = {0140-6736},
+doi = {https://doi.org/10.1016/S0140-6736(14)61122-3},
+url = {https://www.sciencedirect.com/science/article/pii/S0140673614611223},
+author = {Mahesh K B Parmar and James Carpenter and Matthew R Sydes}
+}
 
+@Article{Genz2009,
+  title = {Computation of Multivariate Normal and t Probabilities},
+  author = {Alan Genz and Frank Bretz},
+  series = {Lecture Notes in Statistics},
+  year = {2009},
+  publisher = {Springer-Verlag},
+  address = {Heidelberg},
+  isbn = {978-3-642-01688-2},
+}
