### 2024-04-13 Tpatni719/gsMAMS#17 @Tpatni719
Using \ inside a sentence is not something I would expect in a standard of writing in a journal. I suggest replacing instances of this with the work "or". E.g., "Based on the simulation results, the probability of success (or power)..."
Additionally, using words like "around" here is informal, and I would replace instances of "around" with "approximately", or remove the word entirely.
>>> Done!

The authors should provide a citation for this statement, and should state why they are inefficient.
>>> I have provided the citation and I think citation is sufficient here.

The authors should state what the computation effort here means. What is high? Is it 5 minutes, 10 minutes? A day? Why is this a problem? Is this the only problem that this package, gsMAMS is solving?
>>> I have quantified the high computational effort of MAMS package relative to our package in the computational aspects.
And a clinician/researcher doesn't want to wait for longer duration(e.g. 3 hours) just to get the design parameters and what if he/she wants to tweak some parameters to see how the design is changing then again, the clinician has to wait for 3hours to get the design parameters. So, this is a very obvious problem which I think I don't have to explain it explicitly for the target audience.

I would suggest stating "The long computational time is a major drawback of using the MAMS package.", rather than stating "hurdle".
>>> Done!!

I would argue that this package has reasonably high computational complexity, as there are many large functions in the package. What do the authors mean by the complexity being very low, and why is this relevant and important?
>>>> Our design functions have low computational time relative to MAMS package which is what we have demonstrated in the paper.  We have not compared the operating characteristics functions. This is relevant and important for the reasons mentioned above and just to add an additional point, sometimes, we have to change the design parameters for a trial which is already running. So, in that case, we can't wait for 4-5 hours just to get the design parameters as we don't want to enroll further patients if the trial claimed futility based on the new parameters.

I suggest describing all acronyms in their first use. What does FWER stand for? What is a Dunnett correction, and why is that important? Could the authors provide a citation here?
>>> Dunnett correction is a common multiple testing comparison procedure and the audience using such package is already cognizant of such methods. And respectfully, I think the current information is sufficient to implement the package and understand the functions. I have mentioned the links of the paper in the description for people interested in the methodology.

The mvtnorm package should be cited here. But more importantly, why are these densities being evaluated in the first place? The authors have provided this sentence with no context for the reasons of this being used.
>>> I have cited the package and these densities are evaluated because under global null, we assume a multivariate normal distribution and again, all the details are in the methodology paper.

  This sentence suggests that there are no limitations, but then states that there are indeed limitations. I suggest clearly stating that it can be only used for up to 10 stages? Is that what the authors mean, and does this mean that there can be 100 arms, and 10 stages?
>>> In an **actual clinical trial** setting, it is very very unlikely to have more than 10  interim looks. So, contextually, it is not a limitation considering the nature of a clinical trial and that's why we have structured the paragraph in such a way.

This sentence should be rewritten for expression - e.g.,
This result should then be references with a table of computation time. I would suggest using the microbenchmark or bench R packages to demonstrate the comparison.
>>> I have changed the wording of the expression and I have already provided an example in the paper which I think is enough to demonstrate the low computational time of our package.

When mentioning functions in the text they should go in backticks and parentheses added. e.g., deign_cont(). Additionally there is an errant } after argm(delta0}. Function arguments should be referenced in either backticks or quotes, e.g., frac or "frac". Also, there should be a space between the word and the parentheses.
All functions in the paper should be styled to have spaced around the = and all functions should be written with new lines for each argument, as specified in the documentation, e.g., https://github.com/Tpatni719/gsMAMS/blob/main/R/design_ord.R#L12-L18. This is important because it makes the function easier to read for the user.
>>> Done!!

I'm not sure why "For FWER and Stagewise FWER:" has a column at the end, this should either be a new heading or form some part of the introduction sentence for this paragraph. Functions should also be wrapped in backticks and have parentheses added as mentioned above. The SCPRT function is also mentioned here in all capitals but no example is given and no definition is given of what this function is or means in this context.
>>> FWER and Stagewise FWER are just a sub-heading under each type of outcome. I have addressed the latter part of the function.

The parameters I believe are now all lower case. Also the documentation states that k is the "Number of treatment arms." So should this be k = 5?
>>> I have changed it to lowercase and it is k=4. k is without the control arm.

Why is "group()", and "control()" written like this?
>>> I have checked the paper and I have provided the arguments inside the brackets.

Could the authors indicate which of these parameters link back to the function arguments?
>>> k=4, hr0=1, hr1=0.67
