### 2024-04-13 b10bdfb
Added clinical trial citations
