+++ paper/paper.bib
@@ -58,3 +58,30 @@ author = {Mahesh K B Parmar and James Carpenter and Matthew R Sydes}
   address = {Heidelberg},
   isbn = {978-3-642-01688-2},
 }
+
+@Article{pushpakom2015,
+  title={Telmisartan and Insulin Resistance in HIV (TAILoR): protocol for a dose-ranging phase II randomised open-labelled trial of telmisartan as a strategy for the reduction of insulin resistance in HIV-positive individuals on combination antiretroviral therapy},
+  author={Pushpakom, SP and Taylor, C and Kolamunnage-Dona, R and Spowart, C and Vora, J and García-Fiñana, M and Kemp, GJ and Whitehead, J and Jaki, T and Khoo, S and Williamson, P and Pirmohamed, M},
+  journal={BMJ Open},
+  volume={5},
+  number={10},
+  pages={e009566},
+  year={2015},
+  doi={10.1136/bmjopen-2015-009566},
+  PMID={26474943},
+  PMCID={PMC4611177}
+}
+
+@Article{John1993,
+author = {John Whitehead},
+title ={Application of Sequential Methods to a Phase III Clinical Trial in Stroke},
+journal = {Drug Information Journal},
+volume = {27},
+number = {3},
+pages = {733-740},
+year = {1993},
+doi = {10.1177/009286159302700315},
+}
+
+
+
