### 2024-04-22 Tpatni719/gsMAMS#17 @njtierney
# Reproducibility

> The numbers are correct. I think you didn't use the seed mentioned in the paper. I have just replicated the results for the survival outcome(power configuration)

My point is that there is manual copying and pasting of results, which is a very easy place to introduce human errors. I suggest using rmarkdown to generate the .md format, as this will help eliminate this problem.

Regarding writing, I stated:

> Indeed, trying to write the syntax below to ensure the results were the same as what the author had written led to me being uncertain about where certain parts of these results were referred. For example, the authors state: The overall stopping probability should be around 1 which is the case here. However the stopping probability is listed as:

```
$`Stopping probability under alternative`
 look1  look2
0.3334 0.6666 `
Which is not 1.
```

In response you said:

> These two(0.3334+0.6666 ) add up to 1. I don't know what you mean by "Which is not 1".

My point then is that this is not made clear in the text.

# Writing

> For a clinician who is running a trial, this is pretty much self-explanatory because the main purpose of running this operating characteristics function is to see whether we reach the desired power or not and in this case we did. So, that's why I didn't mention the number exactly and I just mentioned that we reached the desired power which is 90%.

My point is that in a journal standard of writing the writing should be precise. The text could instead state something like what you just said - that "the desired power of 90% has been met, with the power being 91.3%". Does that make sense?

# Expanding on citation

> I have provided the citation and I think citation is sufficient here.

in which the context was my comment on this sentence in the paper:

> Traditional two-arm randomized control trials are not an optimal choice when multiple experimental arms are available for testing efficacy.

I disagree with your comment that the citation is sufficient. My point is this: This sentence does not describe why these are not optimal - with respect to what? Statistical power? Cost? Clinical outcomes? The paper that you reference gives several reasons for why multi arm trials are better, and so I think it is reasonable to add a short sentence describing the reasons they are optimal or a better a choice.

# Computation

> I have quantified the high computational effort of MAMS package relative to our package in the computational aspects.
And a clinician/researcher doesn't want to wait for longer duration(e.g. 3 hours) just to get the design parameters and what if he/she wants to tweak some parameters to see how the design is changing then again, the clinician has to wait for 3hours to get the design parameters. So, this is a very obvious problem which I think I don't have to explain it explicitly for the target audience.

Given that JOSS is a journal focussing on open source software, I think it is reasonable to explain the computational aspects of the software you have written.  You state:

> But the computational effort of obtaining stopping boundaries is very high when the number of stages exceeds 3. The long computational time is the major drawback of using the `MAMS` package.

The phrase "computational effort" is vague. I know that you have given more detail in the "computational aspects" section, but in this specific sentence that I have described, I suggest taking some of what you said above about the long computational time, and putting that into appropriate text in the journal.

# Computation - why is gsMAMS faster

In the paper you have no mentioned **why** your method is so much faster, I think that this is actually really important to address. Is there a special method or approach that you are using? Why is MAMS slow in comparison?

# Writing

> I'm not sure why "For FWER and Stagewise FWER:" has a column at the end, this should either be a new heading or form some part of the introduction sentence for this paragraph.

This still has not been addressed - this should either be a heading, or the start of a sentence.

Thank you making the changes in referencing `k`, `frac`, `hr0` and `hr1`, however there are still some things that need tidying up:

> For survival outcome, we will consider a MAMS trial with five arms (four treatment arms and a control arm, `k`=4) and two interim looks with balanced information time `frac`=c(0.5, 1). The null hazards ratio is (`hr0`)1 and the alternative hazards ratio is (`hr1`)0.67.

The arguments should be in parentheses, e.g., "(`frac` = c(0.5, 1))".
