+++ paper/paper.bib
@@ -83,5 +83,18 @@ year = {1993},
 doi = {10.1177/009286159302700315},
 }
 
+@Article{Li2023,
+author = {Wu, Jianrong and Li, Yimei and Zhu, Liang},
+title = {Group sequential multi-arm multi-stage trial design with treatment selection},
+journal = {Statistics in Medicine},
+volume = {42},
+number = {10},
+pages = {1480-1491},
+doi = {https://doi.org/10.1002/sim.9682},
+url = {https://onlinelibrary.wiley.com/doi/abs/10.1002/sim.9682},
+eprint = {https://onlinelibrary.wiley.com/doi/pdf/10.1002/sim.9682},
+year = {2023}
+}
+
 
 
