+++ R/design_ord.R
@@ -22,12 +22,12 @@ design_ord <- function(alpha, beta, k, prob, or0, or, frac) {
   n <- size_ord(alpha, beta, k=k, prob, or0, or)
   mat <- matrix(NA, nrow = 2, ncol = length(frac))
   rownames(mat) <- c("Cumulative sample size for treatment group", "Cumulative sample size for control group")
-  colnames(mat) <- paste("Stage", 1:length(frac))
+  colnames(mat) <- paste("Stage", seq_len(length(frac)))
   mat[1, ] <- ceiling(n * frac)
   mat[2, ] <- ceiling(n * frac)
   mat1 <- matrix(NA, nrow = 2, ncol = length(frac))
   rownames(mat1) <- c("Lower bound", "Upper bound")
-  colnames(mat1) <- paste("Stage", 1:length(frac))
+  colnames(mat1) <- paste("Stage", seq_len(length(frac)))
   bv <- scprt(alpha = alpha, k = k, frac = frac)
   mat1[1, ] <- bv$lshape
   mat1[2, ] <- bv$ushape
