+++ R/op_power_ord.R
@@ -171,7 +171,7 @@ op_power_ord <- function(alpha, beta, p, frac, or0, or, nsim, prob, seed) {
           p[k[i - 1] + 2] <- z[1, (i)] > z[q, (i)]
 
           g[i, q - 1] <- 0
-          for (o in 1:length(sk)) {
+          for (o in seq_len(length(sk))) {
             g[i, q - 1] <- g[i, q - 1] + prod(p[sk[[o]]])
           }
           next
@@ -190,7 +190,7 @@ op_power_ord <- function(alpha, beta, p, frac, or0, or, nsim, prob, seed) {
         p[k[i - 1] + 2] <- z[1, i] > z[q, i]
 
         g[i, q - 1] <- 0
-        for (o in 1:length(sk)) {
+        for (o in seq_len(length(sk))) {
           g[i, q - 1] <- g[i, q - 1] + prod(p[sk[[o]]])
         }
       }
@@ -249,7 +249,7 @@ op_power_ord <- function(alpha, beta, p, frac, or0, or, nsim, prob, seed) {
     #################################
     lp <- as.data.frame(z)
     z <- list()
-    for (i in 1:length(lp)) {
+    for (i in seq_len(length(lp))) {
       z[[i]] <- lp[, i]
     }
 
