+++ R/op_fwer_ord.R
@@ -161,7 +161,7 @@ op_fwer_ord <- function(alpha, beta, p, frac, or0, or, nsim, prob, seed) {
 
           if (i == j) {
             g[1, q] <- 0
-            for (o in 1:length(sk)) {
+            for (o in seq_len(length(sk))) {
               g[1, q] <- g[1, q] + prod(p[sk[[o]]])
               mp[o, q] <- prod(v[sk[[o]]])
             }
@@ -187,7 +187,7 @@ op_fwer_ord <- function(alpha, beta, p, frac, or0, or, nsim, prob, seed) {
 
         if (i == j) {
           g[1, q] <- 0
-          for (o in 1:length(sk)) {
+          for (o in  seq_len(length(sk))) {
             g[1, q] <- g[1, q] + prod(p[sk[[o]]])
             mp[o, q] <- prod(v[sk[[o]]])
           }
@@ -208,7 +208,7 @@ op_fwer_ord <- function(alpha, beta, p, frac, or0, or, nsim, prob, seed) {
     #################################
     lp <- as.data.frame(z)
     z <- list()
-    for (i in 1:length(lp)) {
+    for (i in  seq_len(length(lp))) {
       z[[i]] <- lp[, i]
     }
 
