### 2024-04-22 Tpatni719/gsMAMS#17 @Tpatni123
My point then is that this is not made clear in the text.
>>> I have changed it but I have already mentioned and explained it thoroughly in the continuous outcome. So, I don't know why this is not clear in survival outcome.
![image](https://github.com/Tpatni719/gsMAMS/assets/42519735/37f810d3-74e0-46c2-95ce-66bae15ab313)

My point is that in a journal standard of writing the writing should be precise. The text could instead state something like what you just said - that "the desired power of 90% has been met, with the power being 91.3%". Does that make sense?
>>>  Done!!
![image](https://github.com/Tpatni719/gsMAMS/assets/42519735/8af0c061-88d1-43ce-891a-2071bcea6f18)

I disagree with your comment that the citation is sufficient. My point is this: This sentence does not describe why these are not optimal - with respect to what? Statistical power? Cost? Clinical outcomes? The paper that you reference gives several reasons for why multi arm trials are better, and so I think it is reasonable to add a short sentence describing the reasons they are optimal or a better a choice.
>>> Done!!
![image](https://github.com/Tpatni719/gsMAMS/assets/42519735/2715caa6-7b28-4ea0-b1d9-8ff124d0be6a)

The phrase "computational effort" is vague. I know that you have given more detail in the "computational aspects" section, but in this specific sentence that I have described, I suggest taking some of what you said above about the long computational time, and putting that into appropriate text in the journal.

>>> Done!!
![image](https://github.com/Tpatni719/gsMAMS/assets/42519735/816a7077-3d37-4119-8ff7-428516b84f2b)

In the paper you have no mentioned why your method is so much faster, I think that this is actually really important to address. Is there a special method or approach that you are using? Why is MAMS slow in comparison?

>>> Done! I have added the necessary details(the details regarding SCPRT boundary calculation is already mentioned in this section) and added the paper for reference.
![image](https://github.com/Tpatni719/gsMAMS/assets/42519735/022798a4-f6e5-4a0b-a5c8-72db707d688c)

This still has not been addressed - this should either be a heading, or the start of a sentence. The arguments should be in parentheses, e.g., "(frac = c(0.5, 1))".
>>> Done!!
![image](https://github.com/Tpatni719/gsMAMS/assets/42519735/54df7d0a-734f-4ef0-87b0-2189b12d7890)
![image](https://github.com/Tpatni719/gsMAMS/assets/42519735/61354658-73c2-483c-83b8-451db9c3d826)
