### 2024-04-23 Tpatni719/gsMAMS#17 @njtierney
Thank you for taking the time to address these changes!

The paper requires a proof read for minor grammar checks - for example there are a few instances of no spaces after parentheses and no spaces after commas. In other journals the paper would go through a proof read from a manuscript editor and they would make these changes. However, I think that JOSS does not provide this, would you be able to check the paper for these changes and other grammatical fixes?

Nearly there!
