### 2024-04-23 Tpatni719/gsMAMS#7 @njtierney
> If you could please fun goodpractice::gp() and return the output when you respond to this that would be great!

Could you run that and paste the output in the comments? Putting the content inside three backticks like so, will render the code neatly:

````
```
[[ paste output of `goodpractice::gp()` here ]]
```
````
