### 2024-04-23 Tpatni719/gsMAMS#7 @Tpatni719
```
── GP gsMAMS ─────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────

It is good practice to

    R/logranktest.R:14:NA
    R/logranktest.R:16:NA
    R/logranktest.R:17:NA
    R/logranktest.R:18:NA
    R/logranktest.R:19:NA
    ... and 483 more lines

  ✖ write short and simple functions. These functions have high cyclomatic complexity (>50): op_power_ord (91), op_power_cont (86),
    op_fwer_ord (82), op_fwer_cont (77), op_power_surv (73), op_fwer_surv (64), size_surv (54). You can make them easier to reason about by
    encapsulating distinct steps of your function into subfunctions.
  ✖ avoid long code lines, it is bad for readability. Also, many people prefer editor windows that are about 80 characters wide. Try
    make your lines shorter than 80 characters

    R\design_cont.R:2:81
    R\design_cont.R:9:81
    R\design_cont.R:11:81
    R\design_cont.R:16:81
    R\design_cont.R:18:81
    ... and 231 more lines

  ✖ not import packages as a whole, as this can cause name clashes between the imported packages. Instead, import only the specific
    functions you need.
──────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────
>
```
