### 2024-04-23 047ba9e
fix various R CMD Check NOTEs
* add paper and LICENSE.md to .Rbuildignore
* use stats::uniroot and stats::punif
