+++ man/design_ord.Rd
@@ -31,8 +31,8 @@ This function generates the design parameters of a clinical trial for ordinal ou
 design_ord(alpha = 0.05,
            beta = 0.1,
            k = 4,
-           prob=c(0.075, 0.182, 0.319, 0.243, 0.015, 0.166),
+           prob = c(0.075, 0.182, 0.319, 0.243, 0.015, 0.166),
            or = 3.06,
            or0 = 1.32,
-           frac=c(1 / 2, 1))
+           frac = c(1 / 2, 1))
 }
