+++ paper/paper.md
@@ -26,7 +26,7 @@ affiliations:
 date:  9 January 2024
 bibliography: paper.bib
 ---
-
+ 
 # Summary
 A multi-arm trial allows simultaneous comparison of multiple experimental treatments with a common control and provides a substantial efficiency advantage compared to conventional randomized controlled trials. A multi-stage trial allows multiple interim looks at the trial outcome so that ineffective arms can be stopped early. In the current R ecosystem, `MAMS` @Jaki2019 is one of the few packages that can handle multiple stages, multiple treatment arms, and different types of endpoints, but the computational effort of obtaining sample size and sequential stopping boundaries is very high when the number of stages exceeds 3. More importantly, their method may lead to an under powered study when the endpoint is time-to-event data. Therefore, we introduce the R package `gsMAMS` (group sequential multi-arm multi-stage (MAMS)) for designing MAMS trials with continuous, ordinal, and survival outcomes, which is computationally very efficient even for a number of stages greater than 3. We also discuss the applications of the package. The `gsMAMS` software facilitates the design and planning of multi-arm clinical trials, which simultaneously compare multiple experimental treatments with a common control group. It is intended to be used by clinical researchers and statisticians so that they can determine appropriate sample sizes and stopping criteria for these trials.
 
