### 2024-04-23 Tpatni719/gsMAMS#17 @Tpatni719
I have done the corrections and thanks again for the recommendations!
