### 2024-04-24 Tpatni719/gsMAMS#17 @njtierney
Great, thanks!
