### 2024-04-29 #6322 @RhysPeploe
Hi, I have one last query, refering back to the below comment regarding commented out sections of code in the functions - I personally would rather have the commented sections removed, however, if @ppxasjsm is happy the code satifies the guidelines then I am happy to tick off

> Hey,
>
> I've notived a fair bit of commented (#) out code in the package, is this something JOSS will want cleaned up before accepting or is this okay? Looking at the guidelines 'JOSS requires that software should be feature-complete (i.e., no half-baked solutions), packaged appropriately according to common community standards for the programming language being used'
>
> ![image](https://private-user-images.githubusercontent.com/68282471/312932838-ca835653-057a-43d0-b637-6d33adfa8dd9.png?jwt=eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpc3MiOiJnaXRodWIuY29tIiwiYXVkIjoicmF3LmdpdGh1YnVzZXJjb250ZW50LmNvbSIsImtleSI6ImtleTUiLCJleHAiOjE3MTQ0MDQzNDIsIm5iZiI6MTcxNDQwNDA0MiwicGF0aCI6Ii82ODI4MjQ3MS8zMTI5MzI4MzgtY2E4MzU2NTMtMDU3YS00M2QwLWI2MzctNmQzM2FkZmE4ZGQ5LnBuZz9YLUFtei1BbGdvcml0aG09QVdTNC1ITUFDLVNIQTI1NiZYLUFtei1DcmVkZW50aWFsPUFLSUFWQ09EWUxTQTUzUFFLNFpBJTJGMjAyNDA0MjklMkZ1cy1lYXN0LTElMkZzMyUyRmF3czRfcmVxdWVzdCZYLUFtei1EYXRlPTIwMjQwNDI5VDE1MjA0MlomWC1BbXotRXhwaXJlcz0zMDAmWC1BbXotU2lnbmF0dXJlPTdlNzA5NWExNWJkMGMwMThhMWFlMDA1MTA2NTE0MWNlMjc5NDcwY2RlMzczNDZkNWRkYzE4N2U3MTY5MGZmNGYmWC1BbXotU2lnbmVkSGVhZGVycz1ob3N0JmFjdG9yX2lkPTAma2V5X2lkPTAmcmVwb19pZD0wIn0.e8ZDmw1cN1tIJ34aLKIEZpyAV3v8T01HqF6oARUuuP8)
