### 2024-04-30 4adc1c3
Update logranktest.R
