### 2024-05-06 #6322 @Tpatni123
Hey,

I've noticed a fair bit of commented (#) out code in the package, is this something JOSS will want cleaned up before accepting or is this okay? Looking at the guidelines 'JOSS requires that software should be feature-complete (i.e., no half-baked solutions), packaged appropriately according to common community standards for the programming language being used'
>>> Hi, @RhysPeploe, can you please check it and tick it off?
