### 2024-05-08 #6322 @RhysPeploe
Hi,

Thanks for removing the commented sections, looks good apart from the below where there is a couple lines still in comments. There's two incidences where there are two lines of comments successively, where one max is needed. I understand some lines can help readability so I am happy for single lines of Hashtags to remain

I will complete my checklist now, on condition that the last remaining commented lines are resolved - thanks

![Screenshot_20240508_121029_Samsung Internet](https://github.com/openjournals/joss-reviews/assets/68282471/287cc1a5-fd49-4f65-b8d5-45c27878d48b)

![Screenshot_20240508_121259_Samsung Internet](https://github.com/openjournals/joss-reviews/assets/68282471/7fcdcf1e-4b16-477e-ab42-6615d5d6be83)
![Screenshot_20240508_121207_Samsung Internet](https://github.com/openjournals/joss-reviews/assets/68282471/50ca4bd3-6458-4753-941e-3533cb5b07ca)
