### 2024-05-08 a9ba122
Update size_surv.R
