### 2024-05-08 1e4c996
Update op_power_surv.R
