### 2024-05-08 fc01910
Update op_power_ord.R
