+++ R/op_power_ord.R
@@ -244,7 +244,7 @@ op_power_ord <- function(alpha, beta, p, frac, or0, or, nsim, prob, seed) {
       }
     }
 
-    #################################
+    
     lp <- as.data.frame(z)
     z <- list()
     for (i in seq_len(length(lp))) {
@@ -301,7 +301,7 @@ op_power_ord <- function(alpha, beta, p, frac, or0, or, nsim, prob, seed) {
         }
       }
     }
-    ############################
+    
     z <- lp
     stopprob <- rep(0, j)
     probfut <- rep(0, j)
