### 2024-05-08 7f78adc
Update op_power_cont.R
