+++ R/op_power_cont.R
@@ -189,7 +189,7 @@ op_power_cont <- function(alpha, beta, p, frac, delta0, delta1, nsim, seed) {
       }
     }
 
-    #################################
+    
     lp <- z
     z <- list()
     for (i in seq_len(length(lp))) {
@@ -246,7 +246,7 @@ op_power_cont <- function(alpha, beta, p, frac, delta0, delta1, nsim, seed) {
         }
       }
     }
-    ############################
+    
     z <- lp
     stopprob <- rep(0, j)
     probfut <- rep(0, j)
