### 2024-05-09 Tpatni719/gsMAMS#19 @ppxasjsm
Small fixes to the manuscript
Hi @Tpatni719, could you have a look at the following two items:

- [x] Fix the first sentence of the summary to reflect you are talking about Clinical Trials here. Maybe something like. In the setting of clinical trials, a multi-armed...
- [x] consider either using gsMAMS package or software, but not use them interchangeably. I would lean towards package.
