### 2024-05-09 #6322 @ppxasjsm
I have gone through the paper and there are two minor things you may consider addressing:
https://github.com/Tpatni719/gsMAMS/issues/19
