### 2024-05-09 #6322 @Tpatni719
> @Tpatni719 could you please check the references, add the missing ones and fix the DOIs where missing.

>>> I have checked the invalid DOIs and they are working. Regarding the missing DOIs, unfortunately, they have not provided DOIs in their BibTex format.
