### 2024-05-09 #6322 @Tpatni719
> Once you have addressed the above, can you make a tagged release and archive, and report the version number and archive DOI in this thread? I can then set the archive version and we can proceed towards acceptance.

Tag: v0.7.2
DOI: https://cran.r-project.org/web/packages/gsMAMS/index.html
(The archiving at CRAN might take 2-3 days but the link will remain the same.)
