+++ DESCRIPTION
@@ -1,6 +1,6 @@
 Package: gsMAMS
 Title: Group Sequential Designs of Multi-Arm Multi-Stage Trials   
-Version: 0.7.1
+Version: 0.7.2
 Imports: mvtnorm, stats, survival 
 Authors@R: c(person("Tushar", "Patni", role = c("aut", "cre"),email = "tushar.patni006@gmail.com"),person("Yimei", "Li", role = "aut"),person("Jianrong", "Wu", role = "aut"))       
 Description: It provides functions to generate operating characteristics and to calculate Sequential Conditional Probability Ratio Tests(SCPRT) efficacy and futility boundary values along with sample/event size of Multi-Arm Multi-Stage(MAMS) trials for different outcomes. The package is based on Jianrong Wu, Yimei Li, Liang Zhu (2023) <doi:10.1002/sim.9682>, Jianrong Wu, Yimei Li (2023) "Group Sequential Multi-Arm Multi-Stage Survival Trial Design with Treatment Selection"(Manuscript accepted for publication) and Jianrong Wu, Yimei Li, Shengping Yang (2023) "Group Sequential Multi-Arm Multi-Stage Trial Design with Ordinal Endpoints"(In preparation).   
