### 2024-05-13 #6322 @ppxasjsm
@Kevin-Mattheus-Moerman is the provided DOI an acceptable DOI? I am not super familiar with R-related things and does not seem like a standard registered DOI.

Also do you have any thoughts on the failed references check. To me they look like we can ignore them?
