+++ paper/paper.bib
@@ -44,7 +44,7 @@ number = {9940},
 pages = {283-284},
 year = {2014},
 issn = {0140-6736},
-doi = {https://doi.org/10.1016/S0140-6736(14)61122-3},
+doi = {10.1016/S0140-6736(14)61122-3},
 url = {https://www.sciencedirect.com/science/article/pii/S0140673614611223},
 author = {Mahesh K B Parmar and James Carpenter and Matthew R Sydes}
 }
@@ -90,7 +90,7 @@ journal = {Statistics in Medicine},
 volume = {42},
 number = {10},
 pages = {1480-1491},
-doi = {https://doi.org/10.1002/sim.9682},
+doi = {10.1002/sim.9682},
 url = {https://onlinelibrary.wiley.com/doi/abs/10.1002/sim.9682},
 eprint = {https://onlinelibrary.wiley.com/doi/pdf/10.1002/sim.9682},
 year = {2023}
