### 2024-05-17 #6322 @njtierney
Searching around I found this article for the package ASD: https://link.springer.com/article/10.1186/1745-6215-12-S1-A8

"Software tools for implementing simulation studies in adaptive seamless designs: introducing R package ASD"

Perhaps there might be other papers for the other pieces of software?
