### 2024-05-17 #6322 @Tpatni719
Thanks but this paper has multiple authors but the package has only single author so I don't know if we can this DOI. I used the citation function and it didn't generate any DOIs for the packages in the bibtex format.
