### 2024-05-17 #6322 @njtierney
@Tpatni123 my point is that if you want to cite the R package ASD, that paper is written by the author, and in the paper it also makes other references to another paper introducing the package: https://link.springer.com/article/10.1186/1745-6215-12-S1-A8

> I used the citation function and it didn't generate any DOIs for the packages in the bibtex format.

It is quite possible that the author of the `asd` package doesn't know how to add the citation format.

It seems like a good-faith move here to try and search for the original papers that reference these papers. What do you think @ppxasjsm ?
