+++ paper/paper.bib
@@ -11,12 +11,15 @@
 
 
 
-@Article{Parsons2016,
-  title = {asd: Simulations for Adaptive Seamless   Designs},
-  author = {Nick Parsons},
-  year = {2016},
-  note = {R package version 2.2},
-  url = {https://CRAN.R-project.org/package=asd},
+@Article{Parsons2011,
+  title={Software tools for implementing simulation studies in adaptive seamless designs: introducing R package ASD},
+  author={Parsons, Nick and Friede, Tim and Todd, Susan and Stallard, Nigel},
+  journal={Trials},
+  volume={12},
+  number={1},
+  pages={A8},
+  year={2011},
+  doi={10.1186/1745-6215-12-S1-A8}
 }
 
 @Article{Hack2022,
