+++ paper/paper.md
@@ -34,7 +34,7 @@ In the field of clinical trial design, a multi-arm trial allows simultaneous com
 # Statement of Need
 Traditional two-arm randomized controlled trials are not an optimal choice when multiple experimental arms are available for testing efficacy (@PARMAR2014283). In such situations, a multiple-arm trial should be preferred, which allows simultaneous comparison of multiple experimental arms with a common control and provides a substantial efficiency advantage. In multi-arm trials, several arms are monitored in a group sequential fashion, with ineffective arms being dropped out of the study. Therefore, multi-arm trials offer a more efficient, cost-effective, and patient-centered approach to clinical research, with the potential to identify superior treatments more reliably than traditional two-arm trials (@PARMAR2014283).
 
-Some packages that are available in R (e.g., `adaptTest` @Vandemeulebroecke2022, `asd` @Hack2022, and `AGSDEST` @Parsons2016) have limitations either in the number of treatment arms that can be incorporated in the package, the number of interim analyses that can be implemented in the package, or the different kinds of outcomes that the package can handle, but the `MAMS` package works well for both multiple treatment arms and multiple stages. It also works for continuous, ordinal, and survival outcomes. However, the computational effort of obtaining stopping boundaries is very high when the number of stages exceeds 3. The long computational time is the major drawback of using the `MAMS` package, and as a result, the process of generating design parameters and evaluating potential modifications can be time-intensive, sometimes taking several hours. This lengthy computation time becomes problematic when researchers or clinicians need to iteratively tweak design parameters and promptly assess the resulting changes in the trial design.
+Some packages that are available in R (e.g., `adaptTest` @Vandemeulebroecke2022, `asd` @Hack2022, and `AGSDEST` @Parsons2011) have limitations either in the number of treatment arms that can be incorporated in the package, the number of interim analyses that can be implemented in the package, or the different kinds of outcomes that the package can handle, but the `MAMS` package works well for both multiple treatment arms and multiple stages. It also works for continuous, ordinal, and survival outcomes. However, the computational effort of obtaining stopping boundaries is very high when the number of stages exceeds 3. The long computational time is the major drawback of using the `MAMS` package, and as a result, the process of generating design parameters and evaluating potential modifications can be time-intensive, sometimes taking several hours. This lengthy computation time becomes problematic when researchers or clinicians need to iteratively tweak design parameters and promptly assess the resulting changes in the trial design.
 
 This paper introduces the R package `gsMAMS` available at https://cran.r-project.org/web/packages/gsMAMS/, which provides functions to obtain sample size, and efficacy and futility boundaries for multiple stages and multiple experimental arms. It also provides functions that generate operating characteristics for designing trials with continuous, ordinal, and survival outcomes. It is computationally very fast compared to the `MAMS` package, even for a number of stages greater than 3.
 
