### 2024-05-17 #6322 @Tpatni719
I tried to find any reference paper for the other two packages and I didn't find any. A lot of papers which referenced these packages also used the same references
![image](https://github.com/openjournals/joss-reviews/assets/13623648/92fcf387-d31d-477f-8e69-5043a59f7e07)
![image](https://github.com/openjournals/joss-reviews/assets/13623648/f7a84e1a-f728-43e3-825e-5899d0706661)
![image](https://github.com/openjournals/joss-reviews/assets/13623648/06132a16-de7e-4ca3-89fd-a60256456e53)
