### 2024-05-25 #6322 @Kevin-Mattheus-Moerman
@ppxasjsm @Tpatni123 unfortunately the CRAN link is not sufficient as it is not a DOI. Please deposit the software on ZENODO and report back here with the DOI.
